Base.@_total_meta — Macro

can be used in place of @assume_effects :total (supposed to be used for bootstrapping)

:total

This setting is the maximum possible set of effects. It currently implies the following other settings:

:consistent

:effect_free

:nothrow

:terminates_globally

:notaskstate

:inaccessiblememonly

:noub

:nortcall

Warning

:total is a very strong assertion and will likely gain additional semantics in future versions of Julia (e.g. if additional effects are added and included in the definition of :total). As a result, it should be used with care. Whenever possible, prefer to use the minimum possible set of specific effect assertions required for a particular application. In cases where a large number of effect overrides apply to a set of functions, a custom macro is recommended over the use of :total.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_terminates_globally_meta — Macro

can be used in place of @assume_effects :terminates_globally (supposed to be used for bootstrapping)

:terminates_globally

The :terminates_globally setting asserts that this method will eventually terminate (either normally or abnormally), i.e. does not loop indefinitely.

Note

This :terminates_globally assertion covers any other methods called by the annotated method.

Note

The compiler will consider this a strong indication that the method will terminate relatively quickly and may (if otherwise legal) call this method at compile time. I.e. it is a bad idea to annotate this setting on a method that technically, but not practically, terminates.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_terminates_globally_noub_meta — Macro

can be used in place of @assume_effects :terminates_globally :noub (supposed to be used for bootstrapping)

:terminates_globally

The :terminates_globally setting asserts that this method will eventually terminate (either normally or abnormally), i.e. does not loop indefinitely.

Note

This :terminates_globally assertion covers any other methods called by the annotated method.

Note

The compiler will consider this a strong indication that the method will terminate relatively quickly and may (if otherwise legal) call this method at compile time. I.e. it is a bad idea to annotate this setting on a method that technically, but not practically, terminates.– doc from julia/base/expr.jl macro assume_effects(args...)

:noub

The :noub setting asserts that the method will not execute any undefined behavior (for any input). Note that undefined behavior may technically cause the method to violate any other effect assertions (such as :consistent or :effect_free) as well, but we do not model this, and they assume the absence of undefined behavior.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_foldable_meta — Macro

can be used in place of @assume_effects :foldable (supposed to be used for bootstrapping)

:foldable

This setting is a convenient shortcut for the set of effects that the compiler requires to be guaranteed to constant fold a call at compile time. It is currently equivalent to the following settings:

:consistent

:effect_free

:terminates_globally

:noub

:nortcall

Note

This list in particular does not include :nothrow. The compiler will still attempt constant propagation and note any thrown error at compile time. Note however, that by the :consistent-cy requirements, any such annotated call must consistently throw given the same argument values.

Note

An explicit @inbounds annotation inside the function will also disable constant folding and not be overridden by :foldable.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_nothrow_noub_meta — Macro

can be used in place of @assume_effects :nothrow :noub (supposed to be used for bootstrapping)

:nothrow

The :nothrow setting asserts that this method does not throw an exception (i.e. will either always return a value or never return).

Note

It is permissible for :nothrow annotated methods to make use of exception handling internally as long as the exception is not rethrown out of the method itself.

Note

If the execution of a method may raise MethodErrors and similar exceptions, then the method is not considered as :nothrow. However, note that environment-dependent errors like StackOverflowError or InterruptException are not modeled by this effect and thus a method that may result in StackOverflowError does not necessarily need to be !:nothrow (although it should usually be !:terminates too).– doc from julia/base/expr.jl macro assume_effects(args...)

:noub

The :noub setting asserts that the method will not execute any undefined behavior (for any input). Note that undefined behavior may technically cause the method to violate any other effect assertions (such as :consistent or :effect_free) as well, but we do not model this, and they assume the absence of undefined behavior.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_terminates_locally_meta — Macro

can be used in place of @assume_effects :terminates_locally (supposed to be used for bootstrapping)

:terminates_locally

The :terminates_locally setting is like :terminates_globally, except that it only applies to syntactic control flow within the annotated method. It is thus a much weaker (and thus safer) assertion that allows for the possibility of non-termination if the method calls some other method that does not terminate.

Note

:terminates_globally implies :terminates_locally.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_terminates_globally_notaskstate_meta — Macro

can be used in place of @assume_effects :terminates_globally :notaskstate (supposed to be used for bootstrapping)

:terminates_globally

The :terminates_globally setting asserts that this method will eventually terminate (either normally or abnormally), i.e. does not loop indefinitely.

Note

This :terminates_globally assertion covers any other methods called by the annotated method.

Note

The compiler will consider this a strong indication that the method will terminate relatively quickly and may (if otherwise legal) call this method at compile time. I.e. it is a bad idea to annotate this setting on a method that technically, but not practically, terminates.– doc from julia/base/expr.jl macro assume_effects(args...)

:notaskstate

The :notaskstate setting asserts that the method does not use or modify the local task state (task local storage, RNG state, etc.) and may thus be safely moved between tasks without observable results.

Note

The implementation of exception handling makes use of state stored in the task object. However, this state is currently not considered to be within the scope of :notaskstate and is tracked separately using the :nothrow effect.

Note

The :notaskstate assertion concerns the state of the currently running task. If a reference to a Task object is obtained by some other means that does not consider which task is currently running, the :notaskstate effect need not be tainted. This is true, even if said task object happens to be === to the currently running task.

Note

Access to task state usually also results in the tainting of other effects, such as :effect_free (if task state is modified) or :consistent (if task state is used in the computation of the result). In particular, code that is not :notaskstate, but is :effect_free and :consistent may still be dead-code-eliminated and thus promoted to :total.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_nospecializeinfer_meta — Macro

Base.@nospecializeinfer function f(args...)
    @nospecialize ...
    ...
end
Base.@nospecializeinfer f(@nospecialize args...) = ...

Tells the compiler to infer f using the declared types of @nospecialized arguments. This can be used to limit the number of compiler-generated specializations during inference.

Examples

julia> f(A::AbstractArray) = g(A)
f (generic function with 1 method)

julia> @noinline Base.@nospecializeinfer g(@nospecialize(A::AbstractArray)) = A[1]
g (generic function with 1 method)

julia> @code_typed f([1.0])
CodeInfo(
1 ─ %1 = invoke Main.g(_2::AbstractArray)::Any
└──      return %1
) => Any

In this example, f will be inferred for each specific type of A, but g will only be inferred once with the declared argument type A::AbstractArray, meaning that the compiler will not likely see the excessive inference time on it while it can not infer the concrete return type of it. Without the @nospecializeinfer, f([1.0]) would infer the return type of g as Float64, indicating that inference ran for g(::Vector{Float64}) despite the prohibition on specialized code generation.

Julia 1.10

Using Base.@nospecializeinfer requires Julia version 1.10.

Base.@_nothrow_meta — Macro

can be used in place of @assume_effects :nothrow (supposed to be used for bootstrapping)

:nothrow

The :nothrow setting asserts that this method does not throw an exception (i.e. will either always return a value or never return).

Note

It is permissible for :nothrow annotated methods to make use of exception handling internally as long as the exception is not rethrown out of the method itself.

Note

If the execution of a method may raise MethodErrors and similar exceptions, then the method is not considered as :nothrow. However, note that environment-dependent errors like StackOverflowError or InterruptException are not modeled by this effect and thus a method that may result in StackOverflowError does not necessarily need to be !:nothrow (although it should usually be !:terminates too).– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_effect_free_terminates_locally_meta — Macro

can be used in place of @assume_effects :effect_free :terminates_locally (supposed to be used for bootstrapping)

:effect_free

The :effect_free setting asserts that the method is free of externally semantically visible side effects. The following is an incomplete list of externally semantically visible side effects:

Changing the value of a global variable.

Mutating the heap (e.g. an array or mutable value), except as noted below

Changing the method table (e.g. through calls to eval)

File/Network/etc. I/O

Task switching

However, the following are explicitly not semantically visible, even if they may be observable:

Memory allocations (both mutable and immutable)

Elapsed time

Garbage collection

Heap mutations of objects whose lifetime does not exceed the method (i.e. were allocated in the method and do not escape).

The returned value (which is externally visible, but not a side effect)

The rule of thumb here is that an externally visible side effect is anything that would affect the execution of the remainder of the program if the function were not executed.

Note

The :effect_free assertion is made both for the method itself and any code that is executed by the method. Keep in mind that the assertion must be valid for all world ages and limit use of this assertion accordingly.– doc from julia/base/expr.jl macro assume_effects(args...)

:terminates_locally

The :terminates_locally setting is like :terminates_globally, except that it only applies to syntactic control flow within the annotated method. It is thus a much weaker (and thus safer) assertion that allows for the possibility of non-termination if the method calls some other method that does not terminate.

Note

:terminates_globally implies :terminates_locally.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_noub_if_noinbounds_meta — Macro

can be used in place of @assume_effects :noub_if_noinbounds (supposed to be used for bootstrapping)

noub::UInt8:

ALWAYS_TRUE: this method is guaranteed to not execute any undefined behavior (for any input).

ALWAYS_FALSE: this method may execute undefined behavior.

NOUB_IF_NOINBOUNDS: this method is guaranteed to not execute any undefined behavior under the assumption that its @boundscheck code is not elided (which happens when the caller does not set nor propagate the @inbounds context)

Note that undefined behavior may technically cause the method to violate any other effect assertions (such as :consistent or :effect_free) as well, but we do not model this, and they assume the absence of undefined behavior.– doc from julia/Compiler/src/effects.jl struct Effects

Base.@_propagate_inbounds_meta — Macro

another version of inlining that propagates an inbounds context @propagate_inbounds

Tells the compiler to inline a function while retaining the caller's inbounds context.

Base.@_noub_meta — Macro

can be used in place of @assume_effects :noub (supposed to be used for bootstrapping)

:noub

The :noub setting asserts that the method will not execute any undefined behavior (for any input). Note that undefined behavior may technically cause the method to violate any other effect assertions (such as :consistent or :effect_free) as well, but we do not model this, and they assume the absence of undefined behavior.– doc from julia/base/expr.jl macro assume_effects(args...)

Base.@_notaskstate_meta — Macro

can be used in place of @assume_effects :notaskstate (supposed to be used for bootstrapping)

:notaskstate

The :notaskstate setting asserts that the method does not use or modify the local task state (task local storage, RNG state, etc.) and may thus be safely moved between tasks without observable results.

Note

The implementation of exception handling makes use of state stored in the task object. However, this state is currently not considered to be within the scope of :notaskstate and is tracked separately using the :nothrow effect.

Note

The :notaskstate assertion concerns the state of the currently running task. If a reference to a Task object is obtained by some other means that does not consider which task is currently running, the :notaskstate effect need not be tainted. This is true, even if said task object happens to be === to the currently running task.

Note

Access to task state usually also results in the tainting of other effects, such as :effect_free (if task state is modified) or :consistent (if task state is used in the computation of the result). In particular, code that is not :notaskstate, but is :effect_free and :consistent may still be dead-code-eliminated and thus promoted to :total.– doc from julia/base/expr.jl macro assume_effects(args...)