API

H3.API

Note

most of the documents take from

https://github.com/uber/h3/tree/master/docs/api

https://github.com/uber/h3/blob/master/src/h3lib/include/h3api.h.in

Types

H3.Lib.H3Index — Type.

H3Index

the H3Index fits within a 64-bit unsigned integer

source

H3.Lib.GeoCoord — Type.

GeoCoord

latitude/longitude in radians

source

H3.Lib.GeoBoundary — Type.

GeoBoundary

cell boundary in latitude/longitude

source

H3.Lib.CoordIJ — Type.

CoordIJ

IJ hexagon coordinates

source

Indexing functions

H3.API.geoToH3 — Function.

geoToH3(location::GeoCoord, resolution::Int)::H3Index

Indexes the location at the specified resolution.

source

H3.API.h3ToGeo — Function.

h3ToGeo(h::H3Index)::GeoCoord

Finds the centroid of the index.

source

H3.API.h3ToGeoBoundary — Function.

h3ToGeoBoundary(h::H3Index)::GeoBoundary

Finds the boundary of the index.

source

Index inspection functions

H3.API.h3GetResolution — Function.

h3GetResolution(h::H3Index)::Int

Returns the resolution of the index.

source

H3.API.h3GetBaseCell — Function.

h3GetBaseCell(h::H3Index)::Int

Returns the base cell number of the index.

source

H3.API.stringToH3 — Function.

stringToH3(str::String)::H3Index

Converts the string representation to H3Index (UInt64) representation.

source

H3.API.h3ToString — Function.

h3ToString(h::H3Index)::String

Converts the H3Index representation of the index to the string representation.

source

H3.API.h3IsValid — Function.

h3IsValid(h::H3Index)::Bool

Returns true if this is a valid H3 index.

source

H3.API.h3IsResClassIII — Function.

h3IsResClassIII(h::H3Index)::Bool

Returns true if this index has a resolution with Class III orientation.

source

H3.API.h3IsPentagon — Function.

h3IsPentagon(h::H3Index)::Bool

Returns true if this index represents a pentagonal cell.

source

Grid traversal functions

H3.API.kRing — Function.

kRing(origin::H3Index, k::Int)::Vector{H3Index}

k-rings produces indices within k distance of the origin index.

source

H3.API.maxKringSize — Function.

maxKringSize(k::Int)::Int

Maximum number of indices that result from the kRing algorithm with the given k.

source

H3.API.kRingDistances — Function.

kRingDistances(origin::H3Index, k::Int)::NamedTuple{(:out, :distances)}

k-rings produces indices within k distance of the origin index.

source

H3.API.hexRange — Function.

hexRange(origin::H3Index, k::Int)::Vector{H3Index}

hexRange produces indexes within k distance of the origin index.

source

H3.API.hexRangeDistances — Function.

hexRangeDistances(origin::H3Index, k::Int)::NamedTuple{(:out, :distances)}

hexRange produces indexes within k distance of the origin index.

source

H3.API.hexRanges — Function.

hexRanges(h3Set::Vector{H3Index}, k::Int)::Vector{H3Index}

hexRanges takes an array of input hex IDs and a max k-ring and returns an array of hexagon IDs sorted first by the original hex IDs and then by the k-ring (0 to max), with no guaranteed sorting within each k-ring group.

source

H3.API.hexRing — Function.

hexRing(origin::H3Index, k::Int)::Vector{H3Index}

Produces the hollow hexagonal ring centered at origin with sides of length k.

source

H3.API.h3Line — Function.

h3Line(origin::H3Index, destination::H3Index)::Vector{H3Index}

Given two H3 indexes, return the line of indexes between them (inclusive).

source

H3.API.h3LineSize — Function.

h3LineSize(origin::H3Index, destination::H3Index)::Int

Number of indexes in a line from the start index to the end index, to be used for allocating memory.

source

H3.API.h3Distance — Function.

h3Distance(origin::H3Index, h::H3Index)::Int

Returns the distance in grid cells between the two indexes.

source

H3.API.experimentalH3ToLocalIj — Function.

experimentalH3ToLocalIj(origin::H3Index, h::H3Index)::CoordIJ

Produces local IJ coordinates for an H3 index anchored by an origin.

source

H3.API.experimentalLocalIjToH3 — Function.

experimentalLocalIjToH3(origin::H3Index, ij::CoordIJ)::H3Index

Produces an H3 index from local IJ coordinates anchored by an origin.

source