H3.Lib

Note

descriptions are taken from

https://github.com/uber/h3/tree/master/docs/api

https://github.com/uber/h3/blob/master/src/h3lib/include/h3api.h.in

Types

H3.Lib.H3Index — Type

const H3Index = UInt64

the H3Index fits within a 64-bit unsigned integer

H3.Lib.GeoCoord — Type

struct GeoCoord
    lat::Cdouble
    lon::Cdouble
end

latitude/longitude in radians

H3.Lib.CoordIJ — Type

struct CoordIJ
    i::Cint
    j::Cint
end

IJ hexagon coordinates

H3.Lib.Vec2d — Type

struct Vec2d
    x::Cdouble
    y::Cdouble
end

2D floating-point vector

H3.Lib.Vec3d — Type

struct Vec3d
    x::Cdouble
    y::Cdouble
    z::Cdouble
end

3D floating point structure

H3.Lib.CoordIJK — Type

struct CoordIJK
    i::Cint
    j::Cint
    k::Cint
end

IJK hexagon coordinates Each axis is spaced 120 degrees apart.

H3.Lib.FaceIJK — Type

struct FaceIJK
    face::Cint
    coord::CoordIJK
end

Face number and ijk coordinates on that face-centered coordinate system

H3.Lib.VertexNode — Type

struct VertexNode
    from::GeoCoord
    to::GeoCoord
    next::Ptr{VertexNode}
end

A single node in a vertex graph, part of a linked list

H3.Lib.VertexGraph — Type

struct VertexGraph
    buckets::Ptr{Ptr{VertexNode}}
    numBuckets::Cint
    size::Cint
    res::Cint
end

A data structure to store a graph of vertices

H3.Lib.Direction — Type

Direction

H3 digit representing ijk+ axes direction. Values will be within the lowest 3 bits of an integer.

H3Index functions

H3.Lib.setH3Index — Function

setH3Index(refh::Ref{H3Index}, res::Int, baseCell::Int, initDigit::Union{Int, Direction})

Initializes an H3 index.

H3.Lib.h3GetIndexDigit — Function

h3GetIndexDigit(h::H3Index, res::Int)::Direction

Gets the resolution res integer digit (0-7) of h3. #define H3_GET_INDEX_DIGIT(h3, res)

H3.Lib.h3GetMode — Function

h3GetMode(h::H3Index)::Int

Gets the integer mode of h3. #define H3_GET_MODE(h3)

H3.Lib.h3SetMode — Function

h3SetMode(refh::Ref{H3Index}, v::Int)

Sets the integer mode of h3 to v. #define H3_SET_MODE(h3, v)

Region functions

H3.Lib.polyfill — Function

polyfill(refpolygon::Ref{GeoPolygon}, res::Int)::Vector{H3Index}

polyfill takes a given GeoJSON-like data structure and preallocated, zeroed memory, and fills it with the hexagons that are contained by the GeoJSON-like data structure.

H3.Lib.maxPolyfillSize — Function

maxPolyfillSize(refpolygon::Ref{GeoPolygon}, res::Int)::Int

maxPolyfillSize returns the number of hexagons to allocate space for when performing a polyfill on the given GeoJSON-like data structure.

H3.Lib.h3SetToLinkedGeo — Function

h3SetToLinkedGeo(h3Set::Vector{H3Index})::Ref{LinkedGeoPolygon}

Create a LinkedGeoPolygon describing the outline(s) of a set of hexagons. Polygon outlines will follow GeoJSON MultiPolygon order: Each polygon will have one outer loop, which is first in the list, followed by any holes.

H3.Lib.destroyLinkedPolygon — Function

destroyLinkedPolygon(refpolygon::Ref{LinkedGeoPolygon})

Free all allocated memory for a linked geo structure. The caller is responsible for freeing memory allocated to the input polygon struct.

Vertex Graph

H3.Lib.initVertexGraph — Function

initVertexGraph(refgraph::Ref{VertexGraph}, numBuckets::Int, res::Int)

Initialize a new VertexGraph

H3.Lib.destroyVertexGraph — Function

destroyVertexGraph(refgraph::Ref{VertexGraph})

Destroy a VertexGraph's sub-objects, freeing their memory. The caller is responsible for freeing memory allocated to the VertexGraph struct itself.