API

API

This page is a dump of all the docstrings found in the code.

CTBase.AbstractOptimalControlModel — Type

abstract type AbstractOptimalControlModel

CTBase.Index — Type

mutable struct Index

Fields

val::Integer

CTBase.OptimalControlModel — Type

mutable struct OptimalControlModel{time_dependence, dimension_usage} <: AbstractOptimalControlModel

Fields

initial_time::Union{Nothing, Real}: Default: nothing

final_time::Union{Nothing, Real}: Default: nothing

time_label::Union{Nothing, String}: Default: nothing

lagrange::Union{Nothing, LagrangeFunction{time_dependence, dimension_usage}} where {time_dependence, dimension_usage}: Default: nothing

mayer::Union{Nothing, MayerFunction{dimension_usage}} where dimension_usage: Default: nothing

criterion::Union{Nothing, Symbol}: Default: nothing

dynamics::Union{Nothing, DynamicsFunction{time_dependence, dimension_usage}} where {time_dependence, dimension_usage}: Default: nothing

dynamics!::Union{Nothing, Function}: Default: nothing

state_dimension::Union{Nothing, Integer}: Default: nothing

state_labels::Vector{String}: Default: Vector{String}()

control_dimension::Union{Nothing, Integer}: Default: nothing

control_labels::Vector{String}: Default: Vector{String}()

constraints::Dict{Symbol, Tuple}: Default: Dict{Symbol, Tuple{Vararg{Any}}}()

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb,
    ub
) -> Tuple{Symbol, Symbol, CTBase.var"#59#60", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}

Add an :initial, :final, :control or :state box constraint (whole range).

Examples

julia> constraint!(ocp, :initial, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :final, [ 0, 0, 0 ], [ 1, 2, 1 ])
julia> constraint!(ocp, :control, [ 0, 0 ], [ 2, 3 ])
julia> constraint!(ocp, :state, [ 0, 0, 0 ], [ 1, 2, 1 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    lb,
    ub
) -> Tuple{Symbol, Symbol, Function, Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, Function, Any, Any}

Add a :boundary, :control, :state or :mixed box functional constraint.

Examples

julia> constraint!(ocp, :boundary, f, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :control, f, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :state, f, [ 0, 0 ], [ 1, 2 ])
julia> constraint!(ocp, :mixed, f, [ 0, 0 ], [ 1, 2 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val
) -> Any
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#59#60", Any, Any}

Add an :initial or :final value constraint on the state.

Examples

julia> constraint!(ocp, :initial, [ 0, 0, 0 ])
julia> constraint!(ocp, :final, [ 0, 0, 0 ])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    val
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}

Add an :initial or :final value constraint on a range of the state.

Examples

julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ])
julia> constraint!(ocp, :final, Index(1), 0)

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    lb,
    ub
) -> Tuple{Symbol, Symbol, CTBase.var"#61#62", Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    rg::Union{Index, UnitRange{<:Integer}},
    lb,
    ub,
    label::Symbol
) -> Tuple{Symbol, Symbol, Any, Any, Any}

Add an :initial, :final, :control or :state box constraint on a range.

Examples

julia> constraint!(ocp, :initial, 2:3, [ 0, 0 ], [1, 2])
julia> constraint!(ocp, :final, Index(1), 0, 2)
julia> constraint!(ocp, :control, Index(1), 0, 2)
julia> constraint!(ocp, :state, 2:3, [ 0, 0 ], [1, 2])

CTBase.constraint! — Function

constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val
) -> Tuple{Symbol, Symbol, Function, Any, Any}
constraint!(
    ocp::OptimalControlModel,
    type::Symbol,
    f::Function,
    val,
    label::Symbol
) -> Tuple{Symbol, Symbol, Function, Any, Any}

Add a :boundary, :control, :state or :mixed value functional constraint.

Examples

julia> constraint!(ocp, :boundary, f, [ 0, 0 ])
julia> constraint!(ocp, :control, f, [ 0, 0 ])
julia> constraint!(ocp, :state, f, [ 0, 0 ])
julia> constraint!(ocp, :mixed, f, [ 0, 0 ])

CTBase.constraint! — Method

constraint!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    f::Function
) -> Any

Provide dynamics (possibly in place).

Examples

julia> constraint!(ocp, :dynamics, f)
julia> constraint!(ocp, :dynamics!, f!)

CTBase.constraint — Method

constraint(ocp::OptimalControlModel, label::Symbol) -> Any

Retrieve a labeled constraint. The result is a function associated with the constraint computation (not taking into account provided value / bounds).

Examples

julia> constraint(ocp, :con)

CTBase.constraint_type — Method

constraint_type(e, t, t0, tf, x, u)

Return the type constraint among :initial, :final, :boundary, :control_range, :control_fun, :state_range, :state_fun, :mixed (:other otherwise), together with the appropriate value (range or updated expression).

Example

julia> t = :t; t0 = 0; tf = :tf; x = :x; u = :u;

julia> constraint_type(:( x[1:2](0) ), t, t0, tf, x, u)
(:initial, 1:2)

julia> constraint_type(:( x[1](0) ), t, t0, tf, x, u)
(:initial, Index(1))

julia> constraint_type(:( 2x[1](0)^2 ), t, t0, tf, x, u)
(:boundary, :(2 * var"x#0"[1] ^ 2))

julia> constraint_type(:( x[1:2](tf) ), t, t0, tf, x, u)
(:final, 1:2)

julia> constraint_type(:( x[1](tf) ), t, t0, tf, x, u)
(:final, Index(1))

julia> constraint_type(:( 2x[1](tf)^2 ), t, t0, tf, x, u)
(:boundary, :(2 * var"x#f"[1] ^ 2))

julia> constraint_type(:( x[1](tf) - x[2](0) ), t, t0, tf, x, u)
(:boundary, :(var"x#f"[1] - var"x#0"[2]))

julia> constraint_type(:( u[1:2](t) ), t, t0, tf, x, u)
(:control_range, 1:2)

julia> constraint_type(:( u[1](t) ), t, t0, tf, x, u)
(:control_range, Index(1))

julia> constraint_type(:( 2u[1](t)^2 ), t, t0, tf, x, u)
(:control_fun, :(2 * u[1] ^ 2))

julia> constraint_type(:( x[1:2](t) ), t, t0, tf, x, u)
(:state_range, 1:2)

julia> constraint_type(:( x[1](t) ), t, t0, tf, x, u)
(:state_range, Index(1))

julia> constraint_type(:( 2x[1](t)^2 ), t, t0, tf, x, u)
(:state_fun, :(2 * x[1] ^ 2))

julia> constraint_type(:( 2u[1](t)^2 * x(t) ), t, t0, tf, x, u)
(:mixed, :((2 * u[1] ^ 2) * x))

julia> constraint_type(:( 2u[1](0)^2 * x(t) ), t, t0, tf, x, u)
:other

CTBase.constraints — Method

constraints(
    ocp::OptimalControlModel
) -> Base.KeySet{Symbol, Dict{Symbol, Tuple}}

Returns a KeySet which contains the list of constraints labels.

Examples

julia> constraints(ocp)

CTBase.control! — Function

control!(
    ocp::OptimalControlModel,
    m::Integer
) -> Vector{String}
control!(
    ocp::OptimalControlModel,
    m::Integer,
    labels::Vector{String}
) -> Vector{String}

Define the control dimension and possibly the names of each coordinate.

Examples

julia> control!(ocp, 2, [ "u₁", "u₂" ])

CTBase.expr_it — Method

expr_it(e, _Expr, f)

Expr iterator: apply _Expr to nodes and f to leaves of the AST.

Example

julia> id(e) = expr_it(e, Expr, x -> x)

CTBase.get — Method

get(sol::UncFreeXfSolution, xx::Union{Symbol, Tuple{Symbol, Integer}})

TBW

CTBase.has — Method

has(e, x, t)

Return true if e contains an x(t), x[i](t) or x[i:j](t) call.

Example

julia> e = :( ∫( x[1](t)^2 + 2*u(t) ) → min )
:(∫((x[1])(t) ^ 2 + 2 * u(t)) → min)

julia> has(e, :x, :t)
true

julia> has(e, :u, :t)
true

CTBase.nlp_constraints — Method

nlp_constraints(
    ocp::OptimalControlModel{time_dependence, dimension_usage}
) -> Tuple{Tuple{Vector{Real}, CTBase.var"#ξ#77", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#η#78", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ψ#79", Vector{Real}}, Tuple{Vector{Real}, CTBase.var"#ϕ#80", Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}, Tuple{Vector{Real}, Vector{Int64}, Vector{Real}}}

Return a 6-tuple of tuples:

(ξl, ξ, ξu) are control constraints

(ηl, η, ηu) are state constraints

(ψl, ψ, ψu) are mixed constraints

(ϕl, ϕ, ϕu) are boundary constraints

(ulb, uind, uub) are control linear constraints of a subset of indices

(xlb, xind, xub) are state linear constraints of a subset of indices

Examples

julia> (ξl, ξ, ξu), (ηl, η, ηu), (ψl, ψ, ψu), (ϕl, ϕ, ϕu), (ulb, uind, uub), (xlb, xind, xub) = nlp_constraints(ocp)

CTBase.objective! — Method

objective!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    f::Function
) -> Any
objective!(
    ocp::OptimalControlModel{time_dependence, dimension_usage},
    type::Symbol,
    f::Function,
    criterion::Symbol
) -> Any

Set the criterion to the function f. Type can be :mayer or :lagrange. Criterion is :min or :max.

Examples

julia> objective!(ocp, (t0, x0, tf, xf) -> tf, :mayer) 

CTBase.remove_constraint! — Method

remove_constraint!(
    ocp::OptimalControlModel,
    label::Symbol
) -> Dict{Symbol, Tuple}

Remove a labeled constraint.

Examples

julia> remove_constraint!(ocp, :con)

CTBase.replace_call — Method

replace_call(e, x, t, y)

Replace calls in e such as x(t), x[i](t) or x[i:j](t) by y, y[i](t) or y[i:j](t), resp.

Example

julia> t = :t; t0 = 0; tf = :tf; x = :x; u = :u;

julia> e = :( x[1](0) * 2x(tf) - x[2](tf) * 2x(0) )
:((x[1])(0) * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

julia> x0 = Symbol(x, 0); e = replace_call(e, x, t0, x0)
:(x0[1] * (2 * x(tf)) - (x[2])(tf) * (2x0))

julia> xf = Symbol(x, "f"); replace_call(ans, x, tf, xf)
:(x0[1] * (2xf) - xf[2] * (2x0))

CTBase.state! — Function

state!(
    ocp::OptimalControlModel,
    n::Integer
) -> Vector{String}
state!(
    ocp::OptimalControlModel,
    n::Integer,
    labels::Vector{String}
) -> Vector{String}

Define the state dimension and possibly the names of each coordinate.

Examples

julia> state!(ocp, 2, [ "x₁", "x₂" ])

CTBase.subs — Method

subs(e, e1, e2)

Substitute expression e1 by expression e2 in expression e.

Examples

julia> e = :( ∫( r(t)^2 + 2u₁(t)) → min )
:(∫(r(t) ^ 2 + 2 * u₁(t)) → min)

julia> subs(e, :r, :( x[1] ))
:(∫((x[1])(t) ^ 2 + 2 * u₁(t)) → min)

julia> e = :( ∫( u₁(t)^2 + 2u₂(t)) → min )
:(∫(u₁(t) ^ 2 + 2 * u₂(t)) → min)

julia> for i ∈ 1:2
       e = subs(e, Symbol(:u, Char(8320+i)), :( u[$i] ))
       end; e
:(∫((u[1])(t) ^ 2 + 2 * (u[2])(t)) → min)

julia> t = :t; t0 = 0; tf = :tf; x = :x; u = :u;

julia> e = :( x[1](0) * 2x(tf) - x[2](tf) * 2x(0) )
:((x[1])(0) * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

julia> x0 = Symbol(x, 0); subs(e, :( $x[1]($(t0)) ), :( $x0[1] ))
:(x0[1] * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

CTBase.time! — Function

time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real}
) -> String
time!(
    ocp::OptimalControlModel,
    times::AbstractVector{<:Real},
    label::String
) -> String

Fix initial and final times to t[1] and t[2], respectively.

Examples

```jldoctest julia> time!(ocp, [ 0, 1 ], "t")

CTBase.time! — Function

time!(
    ocp::OptimalControlModel,
    type::Symbol,
    time::Real
) -> String
time!(
    ocp::OptimalControlModel,
    type::Symbol,
    time::Real,
    label::String
) -> String

Fix initial (resp. final) time, the final (resp. initial) time being variable (free), when type is :initial. And conversely when type is :final.

Examples

```jldoctest julia> time!(ocp, :initial, 0, "t") julia> time!(ocp, :final, 1, "t")