Private functions

Index

CTBase.CTCallback

CTBase.CTCallbacks

CTBase.PrintCallback

CTBase.PrintCallback

CTBase.StopCallback

CTBase.StopCallback

CTBase.get_priority_print_callbacks

CTBase.get_priority_stop_callbacks

Documentation

CTBase.DescVarArg — Constant

A description is a tuple of symbols, that is a Tuple{Vararg{Symbol}}.

CTBase.RangeConstraint — Type

Type alias for an index or range.

CTBase.AbstractCTFunction — Type

abstract type AbstractCTFunction <: Function

Abstract type for functions.

CTBase.AbstractOptimalControlModel — Type

abstract type AbstractOptimalControlModel

CTBase.AbstractOptimalControlSolution — Type

abstract type AbstractOptimalControlSolution

Abstract type for optimal control solutions.

CTBase.AbstractPlotTreeElement — Type

abstract type AbstractPlotTreeElement

Abstract node for plot.

CTBase.AbstractVectorField — Type

abstract type AbstractVectorField{time_dependence, variable_dependence}

Abstract type for vectorfields.

CTBase.ParsingInfo — Type

mutable struct ParsingInfo

Fields

CTBase.PlotLeaf — Type

struct PlotLeaf <: CTBase.AbstractPlotTreeElement

A leaf of a plot tree.

CTBase.PlotNode — Type

struct PlotNode <: CTBase.AbstractPlotTreeElement

A node of a plot tree.

CTBase.PrintCallbacks — Type

Tuple of print callbacks.

CTBase.StopCallbacks — Type

Tuple of stop callbacks.

Base.:\ — Method

\(
    x::Tuple{Vararg{Symbol}},
    y::Tuple{Vararg{Symbol}}
) -> Tuple{Vararg{Symbol}}

Return the difference between the description x and the description y.

Example

julia> (:a, :b) \ (:a,)
(:b,)

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    ocp::OptimalControlModel
)

Print the optimal control problem.

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    sol::OptimalControlSolution
)

Prints the solution.

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    descriptions::Tuple{Vararg{Tuple{Vararg{Symbol}}}}
)

Print a tuple of descriptions.

Base.showerror — Method

showerror(io::IO, e::AmbiguousDescription)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::IncorrectArgument)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::IncorrectMethod)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::IncorrectOutput)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::NotImplemented)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::ParsingError)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::UnauthorizedCall)

Print the exception.

CTBase.:⅋ — Method

⅋(
    X::VectorField{Autonomous, V<:VariableDependence},
    Y::VectorField{Autonomous, V<:VariableDependence}
) -> VectorField{Autonomous}

"Directional derivative" of a vector field: internal and only used to compute efficiently the Lie bracket of two vector fields, autonomous case

Example

julia> X = VectorField(x -> [x[2], -x[1]])
julia> Y = VectorField(x -> [x[1], x[2]])
julia> CTBase.:(⅋)(X, Y)([1, 2])
[2, -1]

CTBase.:⅋ — Method

⅋(
    X::VectorField{NonAutonomous, V<:VariableDependence},
    Y::VectorField{NonAutonomous, V<:VariableDependence}
) -> VectorField{NonAutonomous}

"Directional derivative" of a vector field: internal and only used to compute efficiently the Lie bracket of two vector fields, nonautonomous case

Example

julia> X = VectorField((t, x, v) -> [t + v[1] + v[2] + x[2], -x[1]], NonFixed, NonAutonomous)
julia> Y = VectorField((t, x, v) ->  [v[1] + v[2] + x[1], x[2]], NonFixed, NonAutonomous)
julia> CTBase.:(⅋)(X, Y)(1, [1, 2], [2, 3])
[8, -1]

CTBase.__check_all_set — Method

__check_all_set(
    ocp::OptimalControlModel
) -> Union{Nothing, Bool}

Check if the parameters of an ocp are set.

CTBase.__check_control_set — Method

__check_control_set(ocp::OptimalControlModel) -> Bool

Throw UnauthorizedCall exception if the control of an ocp is not set.

CTBase.__check_dependencies — Method

__check_dependencies(
    dependencies::Tuple{Vararg{DataType}}
) -> Bool

Throw IncorrectArgument exception if dependencies arguments are incorrect.

CTBase.__check_is_time_set — Method

__check_is_time_set(ocp::OptimalControlModel) -> Bool

Throw UnauthorizedCall exception if the time of an ocp is not set.

CTBase.__check_state_set — Method

__check_state_set(ocp::OptimalControlModel) -> Bool

Throw UnauthorizedCall exception if the state of an ocp is not set.

CTBase.__check_variable_set — Method

__check_variable_set(
    ocp::OptimalControlModel{<:TimeDependence, Fixed}
)

Do nothing, no variable for fixed ocp.

CTBase.__check_variable_set — Method

__check_variable_set(
    ocp::OptimalControlModel{<:TimeDependence, NonFixed}
) -> Bool

Throw UnauthorizedCall exception if the variable of an ocp is not set.

CTBase.__get_data_plot — Method

__get_data_plot(
    sol::OptimalControlSolution,
    xx::Union{Symbol, Tuple{Symbol, Integer}}
) -> Any

Get the data for plotting.

CTBase.__initial_plot — Method

__initial_plot(
    sol::OptimalControlSolution;
    layout,
    control,
    kwargs...
) -> Any

Initial plot.

CTBase.__plot_time! — Method

__plot_time!(
    p::Union{Plots.Plot, Plots.Subplot},
    sol::OptimalControlSolution,
    d::Integer,
    s::Symbol;
    t_label,
    labels,
    title,
    kwargs...
)

Update the plot p with a vectorial function of time f(t) ∈ Rᵈ where f is given by the symbol s. The argument s can be :state, :control or :costate.

CTBase.__plot_time! — Method

__plot_time!(
    p::Union{Plots.Plot, Plots.Subplot},
    sol::OptimalControlSolution,
    s::Symbol,
    i::Integer;
    t_label,
    label,
    kwargs...
)

Update the plot p with the i-th component of a vectorial function of time f(t) ∈ Rᵈ where f is given by the symbol s. The argument s can be :state, :control or :costate.

CTBase.__plot_time — Method

__plot_time(
    sol::OptimalControlSolution,
    d::Integer,
    s::Symbol;
    t_label,
    labels,
    title,
    kwargs...
)

Plot a vectorial function of time f(t) ∈ Rᵈ where f is given by the symbol s. The argument s can be :state, :control or :costate.

CTBase.__plot_time — Method

__plot_time(
    sol::OptimalControlSolution,
    s::Symbol,
    i::Integer;
    t_label,
    label,
    kwargs...
)

Plot the i-th component of a vectorial function of time f(t) ∈ Rᵈ where f is given by the symbol s. The argument s can be :state, :control or :costate.

CTBase.__plot_tree — Function

__plot_tree(node::CTBase.PlotNode) -> Any
__plot_tree(
    node::CTBase.PlotNode,
    depth::Integer;
    kwargs...
) -> Any

Plot a node.

CTBase.__plot_tree — Method

__plot_tree(
    leaf::CTBase.PlotLeaf,
    depth::Integer;
    kwargs...
) -> Plots.Plot

Plot a leaf.

CTBase.__width — Method

__width(r::Real) -> Expr

Generate a{r*h} where r is a real number and h is the height of the plot.

CTBase.ctindice — Method

ctindice(i::Integer) -> Char

Return i ∈ [0, 9] as a subscript.

CTBase.ctupperscript — Method

ctupperscript(i::Integer) -> Char

Return i ∈ [0, 9] as an upperscript.

CTBase.expand — Method

expand(x::Matrix{<:Real}) -> Vector{<:Real}

Return expand(matrix2vec(x, 1))

CTBase.expand — Method

expand(x::Vector{<:Real}) -> Vector{<:Real}

Return x.

CTBase.expand — Method

expand(x::Vector{<:Vector{<:Real}}) -> Vector{<:Real}

Equivalent to vec2vec(x)

CTBase.expr_it — Method

expr_it(e, _Expr, f) -> Any

Expr iterator: apply _Expr to nodes and f to leaves of the AST.

Example

julia> id(e) = expr_it(e, Expr, x -> x)

CTBase.has — Method

has(e, x, t) -> Union{Missing, Bool}

Return true if e contains an (...x...)(t) call.

Example

julia> e = :( ∫( x[1](t)^2 + 2*u(t) ) → min )
:(∫((x[1])(t) ^ 2 + 2 * u(t)) → min)

julia> has(e, :x, :t)
true

julia> has(e, :u, :t)
true

CTBase.has — Method

has(e, e1) -> Union{Missing, Bool}

Return true if e contains e1.

Example

julia> e = :( ∫( x[1](t)^2 + 2*u(t) ) → min )
:(∫((x[1])(t) ^ 2 + 2 * u(t)) → min)

julia> has(e, 2)
true

julia> has(e, :x)
true

julia> has(e, :min)
true

julia> has(e, :( x[1](t)^2 ))
true

julia> !has(e, :( x[1](t)^3 ))
true

julia> !has(e, 3)
true

julia> !has(e, :max)
true

julia> has(:x, :x)
true

julia> !has(:x, 2)
true

julia> !has(:x, :y)
true

CTBase.matrix2vec — Function

matrix2vec(x::Matrix{<:Real}) -> Vector{<:Vector{<:Real}}
matrix2vec(
    x::Matrix{<:Real},
    dim::Integer
) -> Vector{<:Vector{<:Real}}

Transforms x to a Vector{<:Vector{<:ctNumber}}.

Note.dim ∈ {1, 2} is the dimension along which the matrix is transformed.

CTBase.parse! — Method

parse!(p, ocp, e; log) -> Union{Expr, LineNumberNode}

Foo

Example

Foo

CTBase.subs — Method

subs(e, e1::Union{Real, Symbol}, e2) -> Any

Substitute expression e1 by expression e2 in expression e.

Examples

julia> e = :( ∫( r(t)^2 + 2u₁(t)) → min )
:(∫(r(t) ^ 2 + 2 * u₁(t)) → min)

julia> subs(e, :r, :( x[1] ))
:(∫((x[1])(t) ^ 2 + 2 * u₁(t)) → min)

julia> e = :( ∫( u₁(t)^2 + 2u₂(t)) → min )
:(∫(u₁(t) ^ 2 + 2 * u₂(t)) → min)

julia> for i ∈ 1:2
       e = subs(e, Symbol(:u, Char(8320+i)), :( u[$i] ))
       end; e
:(∫((u[1])(t) ^ 2 + 2 * (u[2])(t)) → min)

julia> t = :t; t0 = 0; tf = :tf; x = :x; u = :u;

julia> e = :( x[1](0) * 2x(tf) - x[2](tf) * 2x(0) )
:((x[1])(0) * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

julia> x0 = Symbol(x, 0); subs(e, :( $x[1]($(t0)) ), :( $x0[1] ))
:(x0[1] * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

CTBase.vec2vec — Method

vec2vec(
    x::Vector{<:Real},
    n::Integer
) -> Vector{<:Vector{<:Real}}

Transforms x to a Vector{<:Vector{<:ctNumber}}.

CTBase.vec2vec — Method

vec2vec(x::Vector{<:Vector{<:Real}}) -> Vector{<:Real}

Transforms x to a Vector{<:ctNumber}.

RecipesBase.apply_recipe — Method

apply_recipe(
    plotattributes::AbstractDict{Symbol, Any},
    sol::OptimalControlSolution,
    xx::Union{Symbol, Tuple{Symbol, Integer}},
    yy::Union{Symbol, Tuple{Symbol, Integer}}
) -> Vector{RecipesBase.RecipeData}

Return x and y for the plot of the optimal control solution sol corresponding respectively to the argument xx and the argument yy.

Notes.

The argument xx can be :time, :state, :control or :costate.

If xx is :time, then, a label is added to the plot.

The argument yy can be :state, :control or :costate.