Types

CTBase.AbstractHamiltonian — Type

abstract type AbstractHamiltonian{time_dependence, variable_dependence}

Abstract type for hamiltonians.

CTBase.BoundaryConstraint — Type

struct BoundaryConstraint{variable_dependence}

Fields

f::Function

The default value for variable_dependence is Fixed.

Warning

When the state is of dimension 1, consider x0 and xf as a scalar. When the constraint is dimension 1, return a scalar.

Examples

julia> B = BoundaryConstraint((x0, xf) -> [xf[2]-x0[1], 2xf[1]+x0[2]^2]) # variable=false by default
julia> B([0, 0], [1, 1])
[1, 2]
julia> B([0, 0], [1, 1],Real[])
[1, 2]
julia> B = BoundaryConstraint((x0, xf, v) -> [v[3]+xf[2]-x0[1], v[1]-v[2]+2xf[1]+x0[2]^2], variable=true)
julia> B([0, 0], [1, 1], [1, 2, 3])
[4, 1]

CTBase.ControlConstraint — Type

struct ControlConstraint{time_dependence, variable_dependence}

Fields

f::Function

Similar to VectorField in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.ControlLaw — Type

struct ControlLaw{time_dependence, variable_dependence}

Fields

f::Function

Similar to Hamiltonian in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.Dynamics — Type

struct Dynamics{time_dependence, variable_dependence}

Fields

f::Function

Similar to Lagrange, but the function f is assumed to return a vector of the same dimension as the state x.

CTBase.FeedbackControl — Type

struct FeedbackControl{time_dependence, variable_dependence}

Fields

f::Function

Similar to VectorField in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.Hamiltonian — Type

struct Hamiltonian{time_dependence, variable_dependence} <: AbstractHamiltonian{time_dependence, variable_dependence}

Fields

f::Function

The default values for time_dependence and variable_dependence are Autonomous and Fixed respectively.

Warning

When the state and costate are of dimension 1, consider x and p as scalars.

Examples

julia> Hamiltonian((x, p) -> x + p, Int64)
IncorrectArgument 
julia> Hamiltonian((x, p) -> x + p, Int64)
IncorrectArgument
julia> H = Hamiltonian((x, p) -> [x[1]^2+2p[2]]) # autonomous=true, variable=false
julia> H([1, 0], [0, 1])
MethodError # H must return a scalar
julia> H = Hamiltonian((x, p) -> x[1]^2+2p[2])
julia> H([1, 0], [0, 1])
3
julia> t = 1
julia> v = Real[]
julia> H(t, [1, 0], [0, 1])
MethodError
julia> H([1, 0], [0, 1], v)
MethodError 
julia> H(t, [1, 0], [0, 1], v)
3
julia> H = Hamiltonian((x, p, v) -> [x[1]^2+2p[2]+v[3]], variable=true)
julia> H([1, 0], [0, 1], [1, 2, 3])
6
julia> H(t, [1, 0], [0, 1], [1, 2, 3])
6
julia> H = Hamiltonian((t, x, p) -> [t+x[1]^2+2p[2]], autonomous=false)
julia> H(1, [1, 0], [0, 1])
4
julia> H(1, [1, 0], [0, 1], v)
4
julia> H = Hamiltonian((t, x, p, v) -> [t+x[1]^2+2p[2]+v[3]], autonomous=false, variable=true)
julia> H(1, [1, 0], [0, 1], [1, 2, 3])
7

CTBase.HamiltonianLift — Type

struct HamiltonianLift{time_dependence, variable_dependence} <: AbstractHamiltonian{time_dependence, variable_dependence}

Lifts

X::VectorField

The values for time_dependence and variable_dependence are deternimed by the values of those for the VectorField.

Warning

When the state and costate are of dimension 1, consider x and p as scalars.

Examples

julia> HamiltonianLift(HamiltonianLift(VectorField(x -> [x[1]^2, 2x[2]], Int64))
IncorrectArgument 
julia> H = HamiltonianLift(VectorField(x -> [x[1]^2, 2x[2]]))
julia> H([1, 2], [1, 1])
5
julia> t = 1
julia> v = Real[]
julia> H(t, [1, 0], [0, 1])
MethodError
julia> H([1, 0], [0, 1], v)
MethodError 
julia> H(t, [1, 0], [0, 1], v)
5
julia> H = HamiltonianLift(VectorField((x, v) -> [x[1]^2, 2x[2]+v[3]], variable=true))
julia> H([1, 0], [0, 1], [1, 2, 3])
3
julia> H(t, [1, 0], [0, 1], [1, 2, 3])
3
julia> H = HamiltonianLift(VectorField((t, x) -> [t+x[1]^2, 2x[2]], autonomous=false))
julia> H(1, [1, 2], [1, 1])
6
julia> H(1, [1, 0], [0, 1], v)
6
julia> H = HamiltonianLift(VectorField((t, x, v) -> [t+x[1]^2, 2x[2]+v[3]], autonomous=false, variable=true))
julia> H(1, [1, 0], [0, 1], [1, 2, 3])
3

CTBase.HamiltonianVectorField — Type

struct HamiltonianVectorField{time_dependence, variable_dependence} <: CTBase.AbstractVectorField{time_dependence, variable_dependence}

Fields

f::Function

The default values for time_dependence and variable_dependence are Autonomous and Fixed respectively.

Warning

When the state and costate are of dimension 1, consider x and p as scalars.

Examples

julia> HamiltonianVectorField((x, p) -> [x[1]^2+2p[2], x[2]-3p[2]^2], Int64)
IncorrectArgument
julia> HamiltonianVectorField((x, p) -> [x[1]^2+2p[2], x[2]-3p[2]^2], Int64)
IncorrectArgument
julia> Hv = HamiltonianVectorField((x, p) -> [x[1]^2+2p[2], x[2]-3p[2]^2]) # autonomous=true, variable=false
julia> Hv([1, 0], [0, 1])
[3, -3]
julia> t = 1
julia> v = Real[]
julia> Hv(t, [1, 0], [0, 1])
MethodError
julia> Hv([1, 0], [0, 1], v)
MethodError
julia> Hv(t, [1, 0], [0, 1], v)
[3, -3]
julia> Hv = HamiltonianVectorField((x, p, v) -> [x[1]^2+2p[2]+v[3], x[2]-3p[2]^2+v[4]], variable=true)
julia> Hv([1, 0], [0, 1], [1, 2, 3, 4])
[6, -3]
julia> Hv(t, [1, 0], [0, 1], [1, 2, 3, 4])
[6, -3]
julia> Hv = HamiltonianVectorField((t, x, p) -> [t+x[1]^2+2p[2], x[2]-3p[2]^2], autonomous=false)
julia> Hv(1, [1, 0], [0, 1])
[4, -3]
julia> Hv(1, [1, 0], [0, 1], v)
[4, -3]
julia> Hv = HamiltonianVectorField((t, x, p, v) -> [t+x[1]^2+2p[2]+v[3], x[2]-3p[2]^2+v[4]], autonomous=false, variable=true)
julia> Hv(1, [1, 0], [0, 1], [1, 2, 3, 4])
[7, -3]

CTBase.Index — Type

mutable struct Index

Fields

val::Integer

CTBase.Lagrange — Type

struct Lagrange{time_dependence, variable_dependence}

Fields

f::Function

The default value for time_dependence and variable_dependence are Autonomous and Fixed respectively.

Warning

When the state is of dimension 1, consider x as a scalar. Same for the control.

Examples

julia> Lagrange((x, u) -> 2x[2]-u[1]^2, Int64)
IncorrectArgument
julia> Lagrange((x, u) -> 2x[2]-u[1]^2, Int64)
IncorrectArgument
julia> L = Lagrange((x, u) -> [2x[2]-u[1]^2], autonomous=true, variable=false)
julia> L([1, 0], [1])
MethodError
julia> L = Lagrange((x, u) -> 2x[2]-u[1]^2, autonomous=true, variable=false)
julia> L([1, 0], [1])
-1
julia> t = 1
julia> v = Real[]
julia> L(t, [1, 0], [1])
MethodError
julia> L([1, 0], [1], v)
MethodError
julia> L(t, [1, 0], [1], v)
-1
julia> L = Lagrange((x, u, v) -> 2x[2]-u[1]^2+v[3], autonomous=true, variable=true)
julia> L([1, 0], [1], [1, 2, 3])
2
julia> L(t, [1, 0], [1], [1, 2, 3])
2
julia> L = Lagrange((t, x, u) -> t+2x[2]-u[1]^2, autonomous=false, variable=false)
julia> L(1, [1, 0], [1])
0
julia> L(1, [1, 0], [1], v)
0
julia> L = Lagrange((t, x, u, v) -> t+2x[2]-u[1]^2+v[3], autonomous=false, variable=true)
julia> L(1, [1, 0], [1], [1, 2, 3])
3

CTBase.Mayer — Type

struct Mayer{variable_dependence}

Fields

f::Function

The default value for variable_dependence is Fixed.

Warning

When the state is of dimension 1, consider x0 and xf as a scalar.

Examples

julia> G = Mayer((x0, xf) -> [xf[2]-x0[1]]) # variable=false by default
julia> G([0, 0], [1, 1])
MethodError
julia> G = Mayer((x0, xf) -> xf[2]-x0[1])
julia> G([0, 0], [1, 1])
1
julia> G([0, 0], [1, 1],Real[])
1
julia> G = Mayer((x0, xf, v) -> v[3]+xf[2]-x0[1], variable=true)
julia> G([0, 0], [1, 1], [1, 2, 3])
4

CTBase.MixedConstraint — Type

struct MixedConstraint{time_dependence, variable_dependence}

Fields

f::Function

Similar to Lagrange in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.Multiplier — Type

struct Multiplier{time_dependence, variable_dependence}

Fields

f::Function

Similar to ControlLaw in the usage.

CTBase.OptimalControlModel — Type

mutable struct OptimalControlModel{time_dependence<:TimeDependence, variable_dependence<:VariableDependence} <: CTBase.AbstractOptimalControlModel

Fields

model_expression::Union{Nothing, Expr}: Default: nothing

initial_time::Union{Nothing, Index, Real}: Default: nothing

initial_time_name::Union{Nothing, String}: Default: nothing

final_time::Union{Nothing, Index, Real}: Default: nothing

final_time_name::Union{Nothing, String}: Default: nothing

time_name::Union{Nothing, String}: Default: nothing

control_dimension::Union{Nothing, Integer}: Default: nothing

control_components_names::Union{Nothing, Vector{String}}: Default: nothing

control_name::Union{Nothing, String}: Default: nothing

state_dimension::Union{Nothing, Integer}: Default: nothing

state_components_names::Union{Nothing, Vector{String}}: Default: nothing

state_name::Union{Nothing, String}: Default: nothing

variable_dimension::Union{Nothing, Integer}: Default: nothing

variable_components_names::Union{Nothing, Vector{String}}: Default: nothing

variable_name::Union{Nothing, String}: Default: nothing

lagrange::Union{Nothing, Lagrange}: Default: nothing

mayer::Union{Nothing, Mayer}: Default: nothing

criterion::Union{Nothing, Symbol}: Default: nothing

dynamics::Union{Nothing, Dynamics}: Default: nothing

constraints::Dict{Symbol, Tuple}: Default: Dict{Symbol, Tuple{Vararg{Any}}}()

CTBase.OptimalControlSolution — Type

mutable struct OptimalControlSolution <: CTBase.AbstractOptimalControlSolution

Type of an optimal control solution.

Fields

times::Union{Nothing, StepRangeLen, AbstractVector{<:Real}}: Default: nothing

initial_time_name::Union{Nothing, String}: Default: nothing

final_time_name::Union{Nothing, String}: Default: nothing

time_name::Union{Nothing, String}: Default: nothing

control_dimension::Union{Nothing, Integer}: Default: nothing

control_components_names::Union{Nothing, Vector{String}}: Default: nothing

control_name::Union{Nothing, String}: Default: nothing

control::Union{Nothing, Function}: Default: nothing

state_dimension::Union{Nothing, Integer}: Default: nothing

state_components_names::Union{Nothing, Vector{String}}: Default: nothing

state_name::Union{Nothing, String}: Default: nothing

state::Union{Nothing, Function}: Default: nothing

variable_dimension::Union{Nothing, Integer}: Default: nothing

variable_components_names::Union{Nothing, Vector{String}}: Default: nothing

variable_name::Union{Nothing, String}: Default: nothing

variable::Union{Nothing, Real, AbstractVector{<:Real}}: Default: nothing

costate::Union{Nothing, Function}: Default: nothing

objective::Union{Nothing, Real}: Default: nothing

iterations::Union{Nothing, Integer}: Default: nothing

stopping::Union{Nothing, Symbol}: Default: nothing

message::Union{Nothing, String}: Default: nothing

success::Union{Nothing, Bool}: Default: nothing

infos::Dict{Symbol, Any}: Default: Dict{Symbol, Any}()

CTBase.StateConstraint — Type

struct StateConstraint{time_dependence, variable_dependence}

Fields

f::Function

Similar to VectorField in the usage, but the dimension of the output of the function f is arbitrary.

CTBase.TimeDependence — Type

abstract type TimeDependence

CTBase.VariableConstraint — Type

struct VariableConstraint

Fields

f::Function

Warning

When the variable is of dimension 1, consider v as a scalar.

CTBase.VectorField — Type

struct VectorField{time_dependence, variable_dependence} <: CTBase.AbstractVectorField{time_dependence, variable_dependence}

Fields

f::Function

The default value for time_dependence and variable_dependence are Autonomous and Fixed respectively.

Warning

When the state is of dimension 1, consider x as a scalar.

Examples

julia> VectorField(x -> [x[1]^2, 2x[2]], Int64)
IncorrectArgument
julia> VectorField(x -> [x[1]^2, 2x[2]], Int64)
IncorrectArgument
julia> V = VectorField(x -> [x[1]^2, 2x[2]]) # autonomous=true, variable=false
julia> V([1, -1])
[1, -2]
julia> t = 1
julia> v = Real[]
julia> V(t, [1, -1])
MethodError
julia> V([1, -1], v)
MethodError
julia> V(t, [1, -1], v)
[1, -2]
julia> V = VectorField((x, v) -> [x[1]^2, 2x[2]+v[3]], variable=true)
julia> V([1, -1], [1, 2, 3])
[1, 1]
julia> V(t, [1, -1], [1, 2, 3])
[1, 1]
julia> V = VectorField((t, x) -> [t+x[1]^2, 2x[2]], autonomous=false)
julia> V(1, [1, -1])
[2, -2]
julia> V(1, [1, -1], v)
[2, -2]
julia> V = VectorField((t, x, v) -> [t+x[1]^2, 2x[2]+v[3]], autonomous=false, variable=true)
julia> V(1, [1, -1], [1, 2, 3])
[2, 1]