API

H3.API

Note

the documents taken from

https://github.com/uber/h3/tree/master/docs/api

https://github.com/uber/h3/blob/master/src/h3lib/include/h3api.h.in

Types

H3.Lib.H3Index — Type.

const H3Index = UInt64

the H3Index fits within a 64-bit unsigned integer

source

H3.Lib.GeoCoord — Type.

struct GeoCoord
    lat::Cdouble
    lon::Cdouble
end

latitude/longitude in radians

source

H3.Lib.GeoBoundary — Type.

struct GeoBoundary
    numVerts::Cint
    verts::NTuple{10, GeoCoord}
end

cell boundary in latitude/longitude

source

H3.Lib.CoordIJ — Type.

struct CoordIJ
    i::Cint
    j::Cint
end

IJ hexagon coordinates

source

Indexing functions

H3.API.geoToH3 — Function.

geoToH3(location::GeoCoord, resolution::Int)::H3Index

Indexes the location at the specified resolution.

source

H3.API.h3ToGeo — Function.

h3ToGeo(h::H3Index)::GeoCoord

Finds the centroid of the index.

source

H3.API.h3ToGeoBoundary — Function.

h3ToGeoBoundary(h::H3Index)::Vector{GeoCoord}

Finds the boundary of the index.

source

Index inspection functions

H3.API.h3GetResolution — Function.

h3GetResolution(h::H3Index)::Int

Returns the resolution of the index.

source

H3.API.h3GetBaseCell — Function.

h3GetBaseCell(h::H3Index)::Int

Returns the base cell number of the index.

source

H3.API.stringToH3 — Function.

stringToH3(str::String)::H3Index

Converts the string representation to H3Index (UInt64) representation.

source

H3.API.h3ToString — Function.

h3ToString(h::H3Index)::String

Converts the H3Index representation of the index to the string representation.

source

H3.API.h3IsValid — Function.

h3IsValid(h::H3Index)::Bool

Returns true if this is a valid H3 index.

source

H3.API.h3IsResClassIII — Function.

h3IsResClassIII(h::H3Index)::Bool

Returns true if this index has a resolution with Class III orientation.

source

H3.API.h3IsPentagon — Function.

h3IsPentagon(h::H3Index)::Bool

Returns true if this index represents a pentagonal cell.

source

Grid traversal functions

H3.API.kRing — Function.

kRing(origin::H3Index, k::Int)::Vector{H3Index}

k-rings produces indices within k distance of the origin index.

source

H3.API.maxKringSize — Function.

maxKringSize(k::Int)::Int

Maximum number of indices that result from the kRing algorithm with the given k.

source

H3.API.kRingDistances — Function.

kRingDistances(origin::H3Index, k::Int)::NamedTuple{(:out, :distances)}

k-rings produces indices within k distance of the origin index.

source

H3.API.hexRange — Function.

hexRange(origin::H3Index, k::Int)::Vector{H3Index}

hexRange produces indexes within k distance of the origin index.

source

H3.API.hexRangeDistances — Function.

hexRangeDistances(origin::H3Index, k::Int)::NamedTuple{(:out, :distances)}

hexRange produces indexes within k distance of the origin index.

source

H3.API.hexRanges — Function.

hexRanges(h3Set::Vector{H3Index}, k::Int)::Vector{H3Index}

hexRanges takes an array of input hex IDs and a max k-ring and returns an array of hexagon IDs sorted first by the original hex IDs and then by the k-ring (0 to max), with no guaranteed sorting within each k-ring group.

source

H3.API.hexRing — Function.

hexRing(origin::H3Index, k::Int)::Vector{H3Index}

Produces the hollow hexagonal ring centered at origin with sides of length k.

source

H3.API.h3Line — Function.

h3Line(origin::H3Index, destination::H3Index)::Vector{H3Index}

Given two H3 indexes, return the line of indexes between them (inclusive).

source

H3.API.h3LineSize — Function.

h3LineSize(origin::H3Index, destination::H3Index)::Int

Number of indexes in a line from the start index to the end index, to be used for allocating memory.

source

H3.API.h3Distance — Function.

h3Distance(origin::H3Index, h::H3Index)::Int

Returns the distance in grid cells between the two indexes.

source

H3.API.experimentalH3ToLocalIj — Function.

experimentalH3ToLocalIj(origin::H3Index, h::H3Index)::CoordIJ

Produces local IJ coordinates for an H3 index anchored by an origin.

source

H3.API.experimentalLocalIjToH3 — Function.

experimentalLocalIjToH3(origin::H3Index, ij::CoordIJ)::H3Index

Produces an H3 index from local IJ coordinates anchored by an origin.

source

Hierarchical grid functions

H3.API.h3ToParent — Function.

h3ToParent(h::H3Index, parentRes::Integer)::H3Index

Returns the parent (coarser) index containing h.

source

H3.API.h3ToChildren — Function.

h3ToChildren(h::H3Index, childRes::Integer)::Vector{H3Index}

Populates children with the indexes contained by h at resolution childRes. children must be an array of at least size maxH3ToChildrenSize(h, childRes).

source

H3.API.maxH3ToChildrenSize — Function.

maxH3ToChildrenSize(h::H3Index, childRes::Integer)::Int

Returns the size of the array needed by h3ToChildren for these inputs.

source

H3.API.compact — Function.

compact(h3Set::Vector{H3Index})::Vector{H3Index}

Compacts the set h3Set of indexes as best as possible, into the array compactedSet. compactedSet must be at least the size of h3Set in case the set cannot be compacted.

source

H3.API.uncompact — Function.

uncompact(compactedSet::Vector{H3Index}, res::Int)::Vector{H3Index}

Uncompacts the set compactedSet of indexes to the resolution res.

source

H3.API.maxUncompactSize — Function.

maxUncompactSize(compactedSet::Vector{H3Index}, res::Int)::Int

Returns the size of the array needed by uncompact.

source

Region functions

H3.API.polyfill — Function.

polyfill(geoPolygon::GeoPolygon, res::Int)::Vector{H3Index}

source

H3.API.maxPolyfillSize — Function.

maxPolyfillSize(geoPolygon::GeoPolygon, res::Int)::Int

maxPolyfillSize returns the number of hexagons to allocate space for when performing a polyfill on the given GeoJSON-like data structure.

source

H3.API.h3SetToLinkedGeo — Function.

h3SetToLinkedGeo(h3Set::Vector{H3Index})::Ref{LinkedGeoPolygon}

Create a LinkedGeoPolygon describing the outline(s) of a set of hexagons. Polygon outlines will follow GeoJSON MultiPolygon order: Each polygon will have one outer loop, which is first in the list, followed by any holes.

source

H3.API.destroyLinkedPolygon — Function.

destroyLinkedPolygon(polygon::Ref{LinkedGeoPolygon})

Free all allocated memory for a linked geo structure. The caller is responsible for freeing memory allocated to the input polygon struct.

source

Unidirectional edge functions

H3.API.h3IndexesAreNeighbors — Function.

h3IndexesAreNeighbors(origin::H3Index, destination::H3Index)::Bool

Returns whether or not the provided H3Indexes are neighbors. Returns true if the indexes are neighbors, false otherwise.

source

H3.API.getH3UnidirectionalEdge — Function.

getH3UnidirectionalEdge(origin::H3Index, destination::H3Index)::H3Index

Returns a unidirectional edge H3 index based on the provided origin and destination.

source

H3.API.h3UnidirectionalEdgeIsValid — Function.

h3UnidirectionalEdgeIsValid(edge::H3Index)::Bool

Determines if the provided H3Index is a valid unidirectional edge index. Returns true if it is a unidirectional edge H3Index, otherwise false.

source

H3.API.getOriginH3IndexFromUnidirectionalEdge — Function.

getOriginH3IndexFromUnidirectionalEdge(edge::H3Index)::H3Index

Returns the origin hexagon from the unidirectional edge H3Index.

source

H3.API.getDestinationH3IndexFromUnidirectionalEdge — Function.

getDestinationH3IndexFromUnidirectionalEdge(edge::H3Index)::H3Index

Returns the destination hexagon from the unidirectional edge H3Index.

source

H3.API.getH3IndexesFromUnidirectionalEdge — Function.

getH3IndexesFromUnidirectionalEdge(edge::H3Index)::Tuple{H3Index, H3Index}

Returns the origin, destination pair of hexagon IDs for the given edge ID, which are placed at originDestination[0] and originDestination[1] respectively.

source

H3.API.getH3UnidirectionalEdgesFromHexagon — Function.

getH3UnidirectionalEdgesFromHexagon(origin::H3Index)::Vector{H3Index}

Provides all of the unidirectional edges from the current H3Index. edges must be of length 6, and the number of undirectional edges placed in the array may be less than 6.

source

H3.API.getH3UnidirectionalEdgeBoundary — Function.

getH3UnidirectionalEdgeBoundary(edge::H3Index)::Vector{GeoCoord}

Provides the coordinates defining the unidirectional edge.

source

Miscellaneous H3 functions

H3.API.degsToRads — Function.

degsToRads(degrees::Union{Cdouble,Integer})::Cdouble

Converts degrees to radians.

source

H3.API.radsToDegs — Function.

radsToDegs(radians::Union{Cdouble,Integer})::Cdouble

Converts radians to degrees.

source

H3.API.hexAreaKm2 — Function.

hexAreaKm2(res::Integer)::Cdouble

Average hexagon area in square kilometers at the given resolution.

source

H3.API.hexAreaM2 — Function.

hexAreaM2(res::Integer)::Cdouble

Average hexagon area in square meters at the given resolution.

source

H3.API.edgeLengthKm — Function.

edgeLengthKm(res::Integer)::Cdouble

Average hexagon edge length in kilometers at the given resolution.

source

H3.API.edgeLengthM — Function.

edgeLengthM(res::Integer)::Cdouble

Average hexagon edge length in meters at the given resolution.

source

H3.API.numHexagons — Function.

numHexagons(res::Integer)::Int64

Number of unique H3 indexes at the given resolution.

source

H3.API.getRes0Indexes — Function.

getRes0Indexes()::Vector{H3Index}

All the resolution 0 H3 indexes.

source

H3.API.res0IndexCount — Function.

res0IndexCount()::Cint

Number of resolution 0 H3 indexes.

source