API

H3.API

Note

the descriptions are taken from

https://github.com/uber/h3/tree/master/docs/api

https://github.com/uber/h3/blob/master/src/h3lib/include/h3api.h.in

Indexing functions

H3.API.geoToH3 — Function.

geoToH3(location::GeoCoord, resolution::Int)::H3Index

Indexes the location at the specified resolution.

source

H3.API.h3ToGeo — Function.

h3ToGeo(h::H3Index)::GeoCoord

Finds the centroid of the index.

source

H3.API.h3ToGeoBoundary — Function.

h3ToGeoBoundary(h::H3Index)::Vector{GeoCoord}

Finds the boundary of the index.

source

Index inspection functions

H3.API.h3GetResolution — Function.

h3GetResolution(h::H3Index)::Int

Returns the resolution of the index.

source

H3.API.h3GetBaseCell — Function.

h3GetBaseCell(h::H3Index)::Int

Returns the base cell number of the index.

source

H3.API.stringToH3 — Function.

stringToH3(str::String)::H3Index

Converts the string representation to H3Index (UInt64) representation.

source

H3.API.h3ToString — Function.

h3ToString(h::H3Index)::String

Converts the H3Index representation of the index to the string representation.

source

H3.API.h3IsValid — Function.

h3IsValid(h::H3Index)::Bool

Returns true if this is a valid H3 index.

source

H3.API.h3IsResClassIII — Function.

h3IsResClassIII(h::H3Index)::Bool

Returns true if this index has a resolution with Class III orientation.

source

H3.API.h3IsPentagon — Function.

h3IsPentagon(h::H3Index)::Bool

Returns true if this index represents a pentagonal cell.

source

Grid traversal functions

H3.API.kRing — Function.

kRing(origin::H3Index, k::Int)::Vector{H3Index}

k-rings produces indices within k distance of the origin index.

source

H3.API.maxKringSize — Function.

maxKringSize(k::Int)::Int

Maximum number of indices that result from the kRing algorithm with the given k.

source

H3.API.kRingDistances — Function.

kRingDistances(origin::H3Index, k::Int)::NamedTuple{(:out, :distances)}

k-rings produces indices within k distance of the origin index.

source

H3.API.hexRange — Function.

hexRange(origin::H3Index, k::Int)::Vector{H3Index}

hexRange produces indexes within k distance of the origin index.

source

H3.API.hexRangeDistances — Function.

hexRangeDistances(origin::H3Index, k::Int)::NamedTuple{(:out, :distances)}

hexRange produces indexes within k distance of the origin index.

source

H3.API.hexRanges — Function.

hexRanges(h3Set::Vector{H3Index}, k::Int)::Vector{H3Index}

hexRanges takes an array of input hex IDs and a max k-ring and returns an array of hexagon IDs sorted first by the original hex IDs and then by the k-ring (0 to max), with no guaranteed sorting within each k-ring group.

source

H3.API.hexRing — Function.

hexRing(origin::H3Index, k::Int)::Vector{H3Index}

Produces the hollow hexagonal ring centered at origin with sides of length k.

source

H3.API.h3Line — Function.

h3Line(origin::H3Index, destination::H3Index)::Vector{H3Index}

Given two H3 indexes, return the line of indexes between them (inclusive).

source

H3.API.h3LineSize — Function.

h3LineSize(origin::H3Index, destination::H3Index)::Int

Number of indexes in a line from the start index to the end index, to be used for allocating memory.

source

H3.API.h3Distance — Function.

h3Distance(origin::H3Index, h::H3Index)::Int

Returns the distance in grid cells between the two indexes.

source

H3.API.experimentalH3ToLocalIj — Function.

experimentalH3ToLocalIj(origin::H3Index, h::H3Index)::CoordIJ

Produces local IJ coordinates for an H3 index anchored by an origin.

source

H3.API.experimentalLocalIjToH3 — Function.

experimentalLocalIjToH3(origin::H3Index, ij::CoordIJ)::H3Index

Produces an H3 index from local IJ coordinates anchored by an origin.

source

Hierarchical grid functions

H3.API.h3ToParent — Function.

h3ToParent(h::H3Index, parentRes::Int)::H3Index

Returns the parent (coarser) index containing h.

source

H3.API.h3ToChildren — Function.

h3ToChildren(h::H3Index, childRes::Int)::Vector{H3Index}

Populates children with the indexes contained by h at resolution childRes. children must be an array of at least size maxH3ToChildrenSize(h, childRes).

source

H3.API.maxH3ToChildrenSize — Function.

maxH3ToChildrenSize(h::H3Index, childRes::Int)::Int

Returns the size of the array needed by h3ToChildren for these inputs.

source

H3.API.compact — Function.

compact(h3Set::Vector{H3Index})::Vector{H3Index}

Compacts the set h3Set of indexes as best as possible, into the array compactedSet. compactedSet must be at least the size of h3Set in case the set cannot be compacted.

source

H3.API.uncompact — Function.

uncompact(compactedSet::Vector{H3Index}, res::Int)::Vector{H3Index}

Uncompacts the set compactedSet of indexes to the resolution res.

source

H3.API.maxUncompactSize — Function.

maxUncompactSize(compactedSet::Vector{H3Index}, res::Int)::Int

Returns the size of the array needed by uncompact.

source

Unidirectional edge functions

H3.API.h3IndexesAreNeighbors — Function.

h3IndexesAreNeighbors(origin::H3Index, destination::H3Index)::Bool

Returns whether or not the provided H3Indexes are neighbors. Returns true if the indexes are neighbors, false otherwise.

source

H3.API.getH3UnidirectionalEdge — Function.

getH3UnidirectionalEdge(origin::H3Index, destination::H3Index)::H3Index

Returns a unidirectional edge H3 index based on the provided origin and destination.

source

H3.API.h3UnidirectionalEdgeIsValid — Function.

h3UnidirectionalEdgeIsValid(edge::H3Index)::Bool

Determines if the provided H3Index is a valid unidirectional edge index. Returns true if it is a unidirectional edge H3Index, otherwise false.

source

H3.API.getOriginH3IndexFromUnidirectionalEdge — Function.

getOriginH3IndexFromUnidirectionalEdge(edge::H3Index)::H3Index

Returns the origin hexagon from the unidirectional edge H3Index.

source

H3.API.getDestinationH3IndexFromUnidirectionalEdge — Function.

getDestinationH3IndexFromUnidirectionalEdge(edge::H3Index)::H3Index

Returns the destination hexagon from the unidirectional edge H3Index.

source

H3.API.getH3IndexesFromUnidirectionalEdge — Function.

getH3IndexesFromUnidirectionalEdge(edge::H3Index)::Tuple{H3Index, H3Index}

Returns the origin, destination pair of hexagon IDs for the given edge ID, which are placed at originDestination[0] and originDestination[1] respectively.

source

H3.API.getH3UnidirectionalEdgesFromHexagon — Function.

getH3UnidirectionalEdgesFromHexagon(origin::H3Index)::Vector{H3Index}

Provides all of the unidirectional edges from the current H3Index. edges must be of length 6, and the number of undirectional edges placed in the array may be less than 6.

source

H3.API.getH3UnidirectionalEdgeBoundary — Function.

getH3UnidirectionalEdgeBoundary(edge::H3Index)::Vector{GeoCoord}

Provides the coordinates defining the unidirectional edge.

source

Miscellaneous H3 functions

degsToRads
radsToDegs
hexAreaKm2
hexAreaM2
edgeLengthKm
edgeLengthM
numHexagons
getRes0Indexes
res0IndexCount

Coordinate Systems

H3.API.ijToIjk — Function.

ijToIjk(c::CoordIJ)::CoordIJK

Transforms coordinates from the IJ coordinate system to the IJK+ coordinate system.

source

H3.API.ijkToHex2d — Function.

ijkToHex2d(c::CoordIJK)::Vec2d

Find the center point in 2D cartesian coordinates of a hex.

source

H3.API.ijkToIj — Function.

ijkToIj(c::CoordIJK)::CoordIJ

Transforms coordinates from the IJK+ coordinate system to the IJ coordinate system.

source

H3.API.ijkDistance — Function.

ijkDistance(c1::CoordIJK, c2::CoordIJK)::Int

Finds the distance between the two coordinates. Returns result.

source

H3.API.ijkNormalize — Function.

ijkNormalize(c::CoordIJK)::CoordIJK

Normalizes ijk coordinates by setting the components to the smallest possible values. Works in place.

source

H3.API.h3ToLocalIjk — Function.

h3ToLocalIjk(origin::H3Index, h3::H3Index)::CoordIJK

Produces ijk+ coordinates for an index anchored by an origin. The coordinate space used by this function may have deleted regions or warping due to pentagonal distortion. Coordinates are only comparable if they come from the same origin index. Failure may occur if the index is too far away from the origin or if the index is on the other side of a pentagon.

source

H3.API.h3ToFaceIjk — Function.

h3ToFaceIjk(h::H3Index)::FaceIJK

Convert an H3Index to a FaceIJK address.

source

H3.API.localIjkToH3 — Function.

localIjkToH3(origin::H3Index, ijk::CoordIJK)::H3Index

Produces an index for ijk+ coordinates anchored by an origin. The coordinate space used by this function may have deleted regions or warping due to pentagonal distortion. Failure may occur if the coordinates are too far away from the origin or if the index is on the other side of a pentagon.

source

H3.API.faceIjkToH3 — Function.

faceIjkToH3(faceijk::FaceIJK, res::Int)::H3Index

Convert an FaceIJK address to the corresponding H3Index.

source

H3.API.hex2dToCoordIJK — Function.

hex2dToCoordIJK(v::Vec2d)::CoordIJK

Determine the containing hex in ijk+ coordinates for a 2D cartesian coordinate vector (from DGGRID).

source

H3.API.geoToVec3d — Function.

geoToVec3d(geo::GeoCoord)::Vec3d

Calculate the 3D coordinate on unit sphere from the latitude and longitude.

source

H3.API.geoToFaceIjk — Function.

geoToFaceIjk(geo::GeoCoord, res::Int)::FaceIJK

Encodes a coordinate on the sphere to the FaceIJK address of the containing cell at the specified resolution.

source