CTDirect API

Index

CTDirect.OptimalControlInit

CTDirect.Methods

CTDirect.is_solvable

CTDirect.solve

Available methods

using CTDirect
Methods()

(:adnlp, :ipopt)

Documentation

CTDirect.OptimalControlInit — Type

Initialization of the optimal control problem solution for the NLP problem.

Constructors:

OptimalControlInit(): default initialization

OptimalControlInit(x_init, u_init, v_init): constant vector or function handles

OptimalControlInit(sol): from existing solution

Examples

julia> init = OptimalControlInit()
julia> init = OptimalControlInit(x_init=[0.1, 0.2], u_init=0.3)
julia> init = OptimalControlInit(x_init=[0.1, 0.2], u_init=0.3, v_init=0.5)
julia> init = OptimalControlInit(x_init=[0.1, 0.2], u_init=t->sin(t), v_init=0.5)
julia> init = OptimalControlInit(sol)

CTDirect.Methods — Method

Methods() -> Tuple{Tuple{Vararg{Symbol}}}

Return the list of available methods to solve the optimal control problem.

CTDirect.is_solvable — Method

is_solvable(ocp) -> Bool

Return if the optimal control problem ocp is solvable or not by the method solve.

CTDirect.solve — Method

solve(
    ocp::OptimalControlModel,
    description...;
    display,
    init,
    grid_size,
    print_level,
    mu_strategy,
    kwargs...
) -> OptimalControlSolution

Solve the the optimal control problem ocp by the method given by the (optional) description. Return an OptimalControlSolution from CTBase package, that is an approximation of the optimal solution if the method has converged correctly.

The (optional) description

You can pass a partial description. If you give a partial description, then, if several complete descriptions contains the partial one, then, the method with the highest priority is chosen. The higher in the list, the higher is the priority.

Keyword arguments:

display: print or not information during the resolution

init: an initial condition for the solver

grid_size: number of time steps for the discretization

print_level: print level for the Ipopt solver

mu_strategy: mu strategy for the Ipopt solver

Warning

There is only one available method for the moment: the direct method transforms the optimal control problem into a nonlinear programming problem (NLP) solved by Ipopt, thanks to the package ADNLPModels.

Tip

To see the list of available methods, simply call Methods().

You can pass any other option by a pair keyword=value according to the chosen method. See for instance, Ipopt options.

The default values for the keyword arguments are given here.

julia> solve(ocp)
julia> solve(ocp, :adnlp)
julia> solve(ocp, :adnlp, :ipopt)
julia> solve(ocp, display=false, init=OptimalControlInit(), grid_size=100, print_level=0, mu_strategy="adaptive")