API

API

This page is a dump of all the docstrings found in the code.

CTBase.OptimalControlSolution — Method

OptimalControlSolution(
    ocfs::CTFlows.OptimalControlFlowSolution
) -> CTBase.OptimalControlSolution

Construct an OptimalControlSolution from an OptimalControlFlowSolution.

CTFlows.OptimalControlFlowSolution — Type

struct OptimalControlFlowSolution

Type that represents a solution to an optimal control flow.

Fields

ode_sol::Any

feedback_control::CTBase.ControlLaw

ocp::CTBase.OptimalControlModel

CTFlows.Flow — Method

Flow(
    ocp::CTBase.OptimalControlModel{T, V},
    u_::Function,
    g_::Function,
    μ_::Function;
    alg,
    abstol,
    reltol,
    saveat,
    kwargs_Flow...
) -> CTFlows.OptimalControlFlow{Union{Real, AbstractVector{<:Real}}, Union{Real, AbstractVector{<:Real}}, Real}

Flow from an optimal control problem, a control function in feedback form, a state constraint and its associated multiplier in feedback form.

Example

julia> ocp = Model(time_dependence=:nonautonomous)
julia> f = Flow(ocp, (t, x, p) -> p[1], (t, x) -> x[1] - 1, (t, x, p) -> x[1]+p[1])

Warning

The time dependence of the control function must be consistent with the time dependence of the optimal control problem. The dimension of the output of the control function must be consistent with the dimension usage of the control of the optimal control problem.

CTFlows.Flow — Method

Flow(
    ocp::CTBase.OptimalControlModel{T, V},
    u_::Function;
    alg,
    abstol,
    reltol,
    saveat,
    kwargs_Flow...
) -> CTFlows.OptimalControlFlow{Union{Real, AbstractVector{<:Real}}, Union{Real, AbstractVector{<:Real}}, Real}

Flow from an optimal control problem and a control function in feedback form.

Example

julia> f = Flow(ocp, (x, p) -> p)

Warning

The time dependence of the control function must be consistent with the time dependence of the optimal control problem. The dimension of the output of the control function must be consistent with the dimension usage of the control of the optimal control problem.

CTFlows.__abstol — Method

__abstol() -> Float64

Default absolute tolerance for ODE solvers.

See abstol from OrdinaryDiffEq.

CTFlows.__alg — Method

__alg(

) -> OrdinaryDiffEq.Tsit5{typeof(OrdinaryDiffEq.trivial_limiter!), typeof(OrdinaryDiffEq.trivial_limiter!), Static.False}

Default algorithm for ODE solvers.

See alg from OrdinaryDiffEq.

CTFlows.__reltol — Method

__reltol() -> Float64

Default relative tolerance for ODE solvers.

See reltol from OrdinaryDiffEq.

CTFlows.__saveat — Method

__saveat() -> Vector{Any}

See saveat from OrdinaryDiffEq.

CTFlows.__tstops — Method

__tstops() -> Vector{Real}

See tstops from OrdinaryDiffEq.

CTFlows.hamiltonian_usage — Method

hamiltonian_usage(
    alg,
    abstol,
    reltol,
    saveat;
    kwargs_Flow...
) -> Any

Returns a function that solves ODE problem associated to Hamiltonian vector field.

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw,
    f⁰::CTBase.Lagrange,
    p⁰::Real,
    s::Real,
    g::CTBase.MixedConstraint,
    μ::CTBase.Multiplier
) -> CTFlows.var"#H#32"

Constructs the Hamiltonian:

H(t, x, p) = p ⋅ f(t, x, u(t, x, p)) + s p⁰ f⁰(t, x, u(t, x, p)) + μ(t, x, p) ⋅ g(t, x, u(t, x, p))

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw,
    f⁰::CTBase.Lagrange,
    p⁰::Real,
    s::Real
) -> CTFlows.var"#H#30"

Constructs the Hamiltonian:

H(t, x, p) = p ⋅ f(t, x, u(t, x, p)) + s p⁰ f⁰(t, x, u(t, x, p))

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw,
    g::CTBase.MixedConstraint,
    μ::CTBase.Multiplier
) -> CTFlows.var"#H#31"

Constructs the Hamiltonian:

H(t, x, p) = p ⋅ f(t, x, u(t, x, p)) + μ(t, x, p) ⋅ g(t, x, u(t, x, p))

CTFlows.makeH — Method

makeH(
    f::CTBase.Dynamics,
    u::CTBase.ControlLaw
) -> CTFlows.var"#28#29"

Constructs the Hamiltonian:

H(t, x, p) = p f(t, x, u(t, x, p))

CTFlows.rhs — Method

rhs(h::Hamiltonian) -> CTFlows.var"#rhs!#8"

The right and side from a Hamiltonian.

CTFlows.vector_field_usage — Method

vector_field_usage(
    alg,
    abstol,
    reltol,
    saveat;
    kwargs_Flow...
) -> Any

Returns a function that solves ODE problem associated to classical vector field.