API

CTProblems.CTProblems — Module

CTProblems module.

Lists all the imported modules and packages:

Base

CTBase

CTFlows

Core

DocStringExtensions

MINPACK

List of all the exported names:

Problem

Problems

plot

CTProblems.AbstractCTProblem — Type

abstract type AbstractCTProblem

CTProblems.CTProblemsException — Type

abstract type CTProblemsException <: Exception

CTProblems.NonExistingProblem — Type

struct NonExistingProblem <: CTProblems.CTProblemsException

Fields

example::Tuple{Vararg{Symbol}}

CTProblems.OptimalControlProblem — Type

struct OptimalControlProblem{example} <: CTProblems.AbstractCTProblem

Fields

description::String

model::CTBase.OptimalControlModel

solution::CTBase.OptimalControlSolution

Example

julia> using CTProblems
julia> using CTBase
julia> description = "My empty optimal control problem"
julia> ocp = Model()
julia> sol = OptimalControlSolution()
julia> prob = CTProblems.OptimalControlProblem{(:ocp, :empty)}(description, ocp, sol)
julia> prob isa CTProblems.AbstractCTProblem
true
julia> prob isa CTProblems.OptimalControlProblem
true
julia> prob isa isa CTProblems.OptimalControlProblem{(:ocp, :empty)}
true
julia> prob isa CTProblems.OptimalControlProblem{(:ocp, :empty, :dummy)}
false

CTProblems.OptimalControlProblem — Method

OptimalControlProblem{example}() where {example}

Throw a NonExistingProblem exception if there is no optimal control problem described by example.

Example

julia> CTProblems.OptimalControlProblem{(:ocp, :dummy)}()
ERROR: there is no optimal control problem described by (:ocp, :dummy)

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    prob::CTProblems.OptimalControlProblem
)

Show the description, the model and the solution of the optimal control problem.

julia> using CTProblems
julia> Problem(:integrator)
description     = Double integrator - energy min
model    (Type) = CTBase.OptimalControlModel{:autonomous, :scalar}
solution (Type) = CTBase.OptimalControlSolution

Base.showerror — Method

showerror(io::IO, e::CTProblems.NonExistingProblem)

Print the error message when the optimal control problem described by e.example does not exist.

CTProblems.Problem — Method

Problem(description...) -> Any

Returns the optimal control problem described by description. See Overview of CTProblems.jl and Problems for details.

CTProblems.Problems — Method

Problems()

Print the list of available problems, see List of problems for details.

RecipesBase.plot — Method

plot(sol; kwargs...) -> Plots.Plot

A binding to CTBase.jl plot function.