Default

Index

CTBase.__callbacks

CTBase.__constraint_label

CTBase.__control_components_names

CTBase.__control_name

CTBase.__criterion_type

CTBase.__display

CTBase.__fun_time_dependence

CTBase.__fun_variable_dependence

CTBase.__init_interpolation

CTBase.__matrix_dimension_stock

CTBase.__mu_strategy_ipopt

CTBase.__ocp_time_dependence

CTBase.__ocp_variable_dependence

CTBase.__print_level_ipopt

CTBase.__state_components_names

CTBase.__state_name

CTBase.__time_name

CTBase.__variable_components_names

CTBase.__variable_name

Documentation

CTBase.__callbacks — Method

__callbacks() -> Tuple{}

Used to set the default value of the callbacks argument. The default value is (), which means that no additional callback is given.

CTBase.__constraint_label — Method

__constraint_label() -> Symbol

Used to set the default value of the label of a constraint. A unique value is given to each constraint using the gensym function and prefixing by :unamed.

CTBase.__control_components_names — Method

__control_components_names(m::Integer, name::String) -> Any

Used to set the default value of the names of the controls. The default value is ["u"] for a one dimensional control, and ["u₁", "u₂", ...] for a multi dimensional control.

CTBase.__control_name — Method

__control_name() -> String

Used to set the default value of the names of the control. The default value is "u".

CTBase.__criterion_type — Method

__criterion_type() -> Symbol

Used to set the default value of the type of criterion. Either :min or :max. The default value is :min. The other possible criterion type is :max.

CTBase.__display — Method

__display() -> Bool

Used to set the default value of the display argument. The default value is true, which means that the output is printed during resolution.

CTBase.__fun_time_dependence — Method

__fun_time_dependence() -> Type{Autonomous}

Used to set the default value of the time dependence of the functions.

The default value is Autonomous, which means that the functions are considered time independent. The other possible time dependence is NonAutonomous, which means that the functions are considered time dependent.

CTBase.__fun_variable_dependence — Method

__fun_variable_dependence() -> Type{Fixed}

Used to set the default value of the variable dependence of the functions.

The default value is Fixed, which means that the functions are considered variable independent. The other possible variable dependence is NonFixed, which means that the functions are considered variable dependent.

CTBase.__init_interpolation — Method

__init_interpolation() -> CTBase.var"#13#14"

Used to set the default interpolation function used for initialisation. The default value is Interpolations.linear_interpolation, which means that the initial guess is linearly interpolated.

CTBase.__matrix_dimension_stock — Method

__matrix_dimension_stock() -> Int64

Used to set the default value of the stockage of elements in a matrix. The default value is 1.

CTBase.__mu_strategy_ipopt — Method

__mu_strategy_ipopt() -> String

Used to set the default value of the μ strategy of ipopt for the direct method. The default value is adaptive.

CTBase.__ocp_time_dependence — Method

__ocp_time_dependence() -> Type{Autonomous}

Used to set the default value of the time dependence of the Optimal Control Problem. The default value is Autonomous, which means that the Optimal Control Problem is considered time independent. The other possible time dependence is NonAutonomous, which means that all the functions used to define the Optimal Control Problem are considered time dependent.

CTBase.__ocp_variable_dependence — Method

__ocp_variable_dependence() -> Type{Fixed}

Used to set the default value of the variable dependence of the Optimal Control Problem. The default value is Fixed, which means that the Optimal Control Problem is considered variable independent. The other possible variable dependence is NonFixed, which means that all the functions used to define the Optimal Control Problem are considered variable dependent.

CTBase.__print_level_ipopt — Method

__print_level_ipopt() -> Int64

Used to set the default value of the print level of ipopt for the direct method. The default value is 5.

CTBase.__state_components_names — Method

__state_components_names(n::Integer, name::String) -> Any

Used to set the default value of the names of the states. The default value is ["x"] for a one dimensional state, and ["x₁", "x₂", ...] for a multi dimensional state.

CTBase.__state_name — Method

__state_name() -> String

Used to set the default value of the name of the state. The default value is "x".

CTBase.__time_name — Method

__time_name() -> String

Used to set the default value of the name of the time. The default value is t.

CTBase.__variable_components_names — Method

__variable_components_names(q::Integer, name::String) -> Any

Used to set the default value of the names of the variables. The default value is ["v"] for a one dimensional variable, and ["v₁", "v₂", ...] for a multi dimensional variable.

CTBase.__variable_name — Method

__variable_name() -> String

Used to set the default value of the names of the variables. The default value is "v".