API for developers

Here is the doc of private functions and types.

CTBase.DescVarArg — Constant

A description is a tuple of symbols, that is a Tuple{Vararg{Symbol}}.

CTBase.AbstractCTFunction — Type

abstract type AbstractCTFunction <: Function

Abstract type for functions.

CTBase.AbstractOptimalControlModel — Type

abstract type AbstractOptimalControlModel

CTBase.AbstractOptimalControlSolution — Type

abstract type AbstractOptimalControlSolution

Abstract type for optimal control solutions.

CTBase.AbstractPlotNode — Type

abstract type AbstractPlotNode

Abstract node for plot.

CTBase.Control — Type

Type alias for a control.

CTBase.PlotLeaf — Type

struct PlotLeaf <: CTBase.AbstractPlotNode

A leaf of a plot tree.

CTBase.PlotNode — Type

struct PlotNode <: CTBase.AbstractPlotNode

A node of a plot tree.

CTBase.PrintCallbacks — Type

Tuple of print callbacks.

CTBase.StopCallbacks — Type

Tuple of stop callbacks.

CTBase.SymbolPlot — Type

struct Tuple{Symbol, Integer}

Type alias for a plot element.

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    sol::OptimalControlSolution
)

Prints the solution.

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    descriptions::Tuple{Vararg{Tuple{Vararg{Symbol}}}}
)

Print a tuple of descriptions.

Base.show — Method

show(
    io::IO,
    _::MIME{Symbol("text/plain")},
    ocp::OptimalControlModel{time_dependence, dimension_usage}
)

Print the optimal control problem.

Base.showerror — Method

showerror(io::IO, e::AmbiguousDescription)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::InconsistentArgument)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::IncorrectArgument)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::IncorrectMethod)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::IncorrectOutput)

Print the exception.

Base.showerror — Method

showerror(io::IO, e::NotImplemented)

Print the exception.

CTBase.__abs_tol_direct_shooting — Method

__abs_tol_direct_shooting() -> Float64

Used to set the default value of the absolute tolerance for the stopping criterion for the direct shooting method. The default value is 10 * eps().

CTBase.__callbacks — Method

__callbacks() -> Tuple{}

Used to set the default value of the callbacks argument. The default value is (), which means that no additional callback is given.

CTBase.__constraint_label — Method

__constraint_label() -> Symbol

Used to set the default value of the label of a constraint. A unique value is given to each constraint using the gensym function and prefixing by :unamed.

CTBase.__control_names — Method

__control_names(m::Integer) -> Any

Used to set the default value of the names of the controls. The default value is ["u"] for a one dimensional control, and ["u₁", "u₂", ...] for a multi dimensional control.

CTBase.__criterion_type — Method

__criterion_type() -> Symbol

Used to set the default value of the type of criterion. Either :min or :max. The default value is :min. The other possible criterion type is :max.

CTBase.__display — Method

__display() -> Bool

Used to set the default value of the display argument. The default value is true, which means that the output is printed during resolution.

CTBase.__fun_dimension_usage — Method

__fun_dimension_usage() -> Symbol

Used to set the default value of the dimension usage of the functions.

The default value is :scalar, which means that the usage of the functions is considered scalar. The other possible usage is :vectorial.

Example

If x is for instance a vector of dimension 2 and u a scalar, then the following usage is considered scalar:

f(x, u) = x[1] + u

If x and u are for instance both scalar, then the following usage is considered scalar:

f(x, u) = x + u

If x is for instance a vector of dimension 2 and u a vector of dimension 3, then the following usage is also considered scalar:

f(x, u) = x[1] + u[1]

A vectorial usage is the following. For instance, if x is a vector of dimension 2 and u a scalar, then the following usage is considered vectorial:

f(x, u) = x[1] + u[1]

CTBase.__fun_time_dependence — Method

__fun_time_dependence() -> Symbol

Used to set the default value of the time dependence of the functions.

The default value is :autonomous, which means that the functions are considered time independent. The other possible time dependence is :nonautonomous, which means that the functions are considered time dependent.

CTBase.__grid_size_direct — Method

__grid_size_direct() -> Int64

Used to set the default value of the grid size for the direct method. The default value is 100.

CTBase.__grid_size_direct_shooting — Method

__grid_size_direct_shooting() -> Int64

Used to set the default value of the grid size for the direct shooting method. The default value is 201.

CTBase.__init_interpolation — Method

__init_interpolation() -> CTBase.var"#45#46"

Used to set the default interpolation function used for initialisation. The default value is Interpolations.linear_interpolation, which means that the initial guess is linearly interpolated.

CTBase.__matrix_dimension_stock — Method

__matrix_dimension_stock() -> Int64

Used to set the default value of the stockage of elements in a matrix. The default value is 1.

CTBase.__max_iter_direct_shooting — Method

__max_iter_direct_shooting() -> Int64

Used to set the default value of the maximal number of iterations for the direct shooting method. The default value is 100.

CTBase.__mu_strategy_ipopt — Method

__mu_strategy_ipopt() -> String

Used to set the default value of the μ strategy of ipopt for the direct method. The default value is adaptive.

CTBase.__ocp_dimension_usage — Method

__ocp_dimension_usage() -> Symbol

Used to set the default value of the dimension usage of the Optimal Control Problem. The default value is :scalar, which means that the usage for all the functions used to define the Optimal Control Problem is considered scalar.

CTBase.__ocp_time_dependence — Method

__ocp_time_dependence() -> Symbol

Used to set the default value of the time dependence of the Optimal Control Problem. The default value is :autonomous, which means that the Optimal Control Problem is considered time independent. The other possible time dependence is :nonautonomous, which means that all the functions used to define the Optimal Control Problem are considered time dependent.

CTBase.__opt_tol_direct_shooting — Method

__opt_tol_direct_shooting() -> Float64

Used to set the default value of the optimality relative tolerance for the stopping criterion for the direct shooting method. The default value is 1e-8.

CTBase.__penalty_term_direct_shooting — Method

__penalty_term_direct_shooting() -> Float64

Used to set the default value of the penalty term in front of the final constraint for the direct shooting method. The default value is 1e4.

CTBase.__print_level_ipopt — Method

__print_level_ipopt() -> Int64

Used to set the default value of the print level of ipopt for the direct method. The default value is 5.

CTBase.__stagnation_tol_direct_shooting — Method

__stagnation_tol_direct_shooting() -> Float64

Used to set the default value of the step stagnation relative tolerance for the stopping criterion for the direct shooting method. The default value is 1e-8.

CTBase.__state_names — Method

__state_names(n::Integer) -> Any

Used to set the default value of the names of the states. The default value is ["x"] for a one dimensional state, and ["x₁", "x₂", ...] for a multi dimensional state.

CTBase.__time_name — Method

__time_name() -> String

Used to set the default value of the name of the time. The default value is t.

CTBase._plot_time! — Method

_plot_time!(
    p,
    sol::OptimalControlSolution,
    s::Symbol,
    i::Integer;
    t_label,
    label,
    kwargs...
)

Update the plot p with the i-th component of a vectorial function of time f(t) ∈ Rᵈ where f is given by the symbol s. The argument s can be :state, :control or :adjoint.

CTBase._plot_time — Method

_plot_time(
    sol::OptimalControlSolution,
    d::Integer,
    s::Symbol;
    t_label,
    labels,
    title,
    kwargs...
)

Plot a vectorial function of time f(t) ∈ Rᵈ where f is given by the symbol s. The argument s can be :state, :control or :adjoint.

CTBase._plot_time — Method

_plot_time(
    sol::OptimalControlSolution,
    s::Symbol,
    i::Integer;
    t_label,
    label,
    kwargs...
)

Plot the i-th component of a vectorial function of time f(t) ∈ Rᵈ where f is given by the symbol s. The argument s can be :state, :control or :adjoint.

CTBase.constraint_type — Method

constraint_type(e, t, t0, tf, x, u)

Return the type constraint among :initial, :final, :boundary, :control_range, :control_fun, :state_range, :state_fun, :mixed (:other otherwise), together with the appropriate value (range or updated expression).

Example

julia> t = :t; t0 = 0; tf = :tf; x = :x; u = :u;

julia> constraint_type(:( x[1:2](0) ), t, t0, tf, x, u)
(:initial, 1:2)

julia> constraint_type(:( x[1](0) ), t, t0, tf, x, u)
(:initial, Index(1))

julia> constraint_type(:( 2x[1](0)^2 ), t, t0, tf, x, u)
(:boundary, :(2 * var"x#0"[1] ^ 2))

julia> constraint_type(:( x[1:2](tf) ), t, t0, tf, x, u)
(:final, 1:2)

julia> constraint_type(:( x[1](tf) ), t, t0, tf, x, u)
(:final, Index(1))

julia> constraint_type(:( 2x[1](tf)^2 ), t, t0, tf, x, u)
(:boundary, :(2 * var"x#f"[1] ^ 2))

julia> constraint_type(:( x[1](tf) - x[2](0) ), t, t0, tf, x, u)
(:boundary, :(var"x#f"[1] - var"x#0"[2]))

julia> constraint_type(:( u[1:2](t) ), t, t0, tf, x, u)
(:control_range, 1:2)

julia> constraint_type(:( u[1](t) ), t, t0, tf, x, u)
(:control_range, Index(1))

julia> constraint_type(:( 2u[1](t)^2 ), t, t0, tf, x, u)
(:control_fun, :(2 * u[1] ^ 2))

julia> constraint_type(:( x[1:2](t) ), t, t0, tf, x, u)
(:state_range, 1:2)

julia> constraint_type(:( x[1](t) ), t, t0, tf, x, u)
(:state_range, Index(1))

julia> constraint_type(:( 2x[1](t)^2 ), t, t0, tf, x, u)
(:state_fun, :(2 * x[1] ^ 2))

julia> constraint_type(:( 2u[1](t)^2 * x(t) ), t, t0, tf, x, u)
(:mixed, :((2 * u[1] ^ 2) * x))

julia> constraint_type(:( 2u[1](0)^2 * x(t) ), t, t0, tf, x, u)
:other

CTBase.ctindice — Method

ctindice(i::Integer) -> Char

Returns i ∈ [0, 9] as a subscript.

CTBase.ctupperscript — Method

ctupperscript(i::Integer) -> Char

Returns i ∈ [0, 9] as an upperscript.

CTBase.expand — Method

expand(x::Matrix{<:Real}) -> Vector{<:Real}

Returns expand(matrix2vec(x, 1))

CTBase.expand — Method

expand(x::Vector{<:Real}) -> Vector{<:Real}

Returns x.

CTBase.expand — Method

expand(x::Vector{<:Vector{<:Real}}) -> Vector{<:Real}

Equivalent to vec2vec(x)

CTBase.expr_it — Method

expr_it(e, _Expr, f)

Expr iterator: apply _Expr to nodes and f to leaves of the AST.

Example

julia> id(e) = expr_it(e, Expr, x -> x)

CTBase.get — Method

get(
    sol::OptimalControlSolution,
    xx::Union{Symbol, Tuple{Symbol, Integer}}
) -> Any

Get the data for plotting.

CTBase.has — Method

has(e, x, t)

Return true if e contains an x(t), x[i](t) or x[i:j](t) call.

Example

julia> e = :( ∫( x[1](t)^2 + 2*u(t) ) → min )
:(∫((x[1])(t) ^ 2 + 2 * u(t)) → min)

julia> has(e, :x, :t)
true

julia> has(e, :u, :t)
true

CTBase.matrix2vec — Function

matrix2vec(x::Matrix{<:Real}) -> Vector{<:Vector{<:Real}}
matrix2vec(
    x::Matrix{<:Real},
    dim::Integer
) -> Vector{<:Vector{<:Real}}

Transforms x to a Vector{<:Vector{<:MyNumber}}.

Note.dim ∈ {1, 2} is the dimension along which the matrix is transformed.

CTBase.replace_call — Method

replace_call(e, x, t, y)

Replace calls in e such as x(t), x[i](t) or x[i:j](t) by y, y[i](t) or y[i:j](t), resp.

Example

julia> t = :t; t0 = 0; tf = :tf; x = :x; u = :u;

julia> e = :( x[1](0) * 2x(tf) - x[2](tf) * 2x(0) )
:((x[1])(0) * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

julia> x0 = Symbol(x, 0); e = replace_call(e, x, t0, x0)
:(x0[1] * (2 * x(tf)) - (x[2])(tf) * (2x0))

julia> xf = Symbol(x, "f"); replace_call(ans, x, tf, xf)
:(x0[1] * (2xf) - xf[2] * (2x0))

CTBase.subs — Method

subs(e, e1, e2)

Substitute expression e1 by expression e2 in expression e.

Examples

julia> e = :( ∫( r(t)^2 + 2u₁(t)) → min )
:(∫(r(t) ^ 2 + 2 * u₁(t)) → min)

julia> subs(e, :r, :( x[1] ))
:(∫((x[1])(t) ^ 2 + 2 * u₁(t)) → min)

julia> e = :( ∫( u₁(t)^2 + 2u₂(t)) → min )
:(∫(u₁(t) ^ 2 + 2 * u₂(t)) → min)

julia> for i ∈ 1:2
       e = subs(e, Symbol(:u, Char(8320+i)), :( u[$i] ))
       end; e
:(∫((u[1])(t) ^ 2 + 2 * (u[2])(t)) → min)

julia> t = :t; t0 = 0; tf = :tf; x = :x; u = :u;

julia> e = :( x[1](0) * 2x(tf) - x[2](tf) * 2x(0) )
:((x[1])(0) * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

julia> x0 = Symbol(x, 0); subs(e, :( $x[1]($(t0)) ), :( $x0[1] ))
:(x0[1] * (2 * x(tf)) - (x[2])(tf) * (2 * x(0)))

CTBase.vec2vec — Method

vec2vec(
    x::Vector{<:Real},
    n::Integer
) -> Vector{<:Vector{<:Real}}

Transforms x to a Vector{<:Vector{<:MyNumber}}.

CTBase.vec2vec — Method

vec2vec(x::Vector{<:Vector{<:Real}}) -> Vector{<:Real}

Transforms x to a Vector{<:MyNumber}.

RecipesBase.apply_recipe — Method

apply_recipe(
    plotattributes::AbstractDict{Symbol, Any},
    sol::OptimalControlSolution,
    xx::Union{Symbol, Tuple{Symbol, Integer}},
    yy::Union{Symbol, Tuple{Symbol, Integer}}
) -> Vector{RecipesBase.RecipeData}

Returns x and y for the plot of the optimal control solution sol corresponding respectively to the argument xx and the argument yy.

Notes.

The argument xx can be :time, :state, :control or :adjoint.

If xx is :time, then, a label is added to the plot.

The argument yy can be :state, :control or :adjoint.

CTBase.@ctfunction_sv — Macro

Generate callable types, that is functions, handling the dimension usage.

Note: This macro is used to generate callable types for functions that are not time dependent. If the function is time dependent, use the macro @ctfunction_td_sv instead.

CTBase.@ctfunction_td_sv — Macro

Generate callable types, that is functions, handling the time dependence and the dimension usages.